\name{affine.fromFunction}
\alias{affine.fromFunction}
\title{affine...}
\usage{\method{affine}{fromFunction} (X, mat, ...)}
\description{affine.fromFunction function}
\details{An affine transformation of an object of class \code{fromFunction}}
\value{the object acted on by the transformation matrix}
\arguments{\item{X}{an object of class fromFunction}
\item{mat}{matrix of affine transformation}
\item{...}{additional arguments}}
