\name{affine.SpatialPolygonsDataFrame}
\alias{affine.SpatialPolygonsDataFrame}
\title{affine...}
\usage{\method{affine}{SpatialPolygonsDataFrame} (X, mat, ...)}
\description{affine.SpatialPolygonsDataFrame function}
\details{An affine transformation of an object of class \code{SpatialPolygonsDataFrame}}
\value{the object acted on by the transformation matrix}
\arguments{\item{X}{an object of class fromFunction}
\item{mat}{matrix of affine transformation}
\item{...}{additional arguments}}
