\name{affine.fromSPDF}
\alias{affine.fromSPDF}
\title{affine...}
\usage{\method{affine}{fromSPDF} (X, mat, ...)}
\description{affine.fromSPDF function}
\details{An affine transformation of an object of class \code{fromSPDF}}
\value{the object acted on by the transformation matrix}
\arguments{\item{X}{an object of class fromSPDF}
\item{mat}{matrix of affine transformation}
\item{...}{additional arguments}}
