\name{affine.fromXYZ}
\alias{affine.fromXYZ}
\title{affine...}
\usage{\method{affine}{fromXYZ} (X, mat, ...)}
\description{affine.fromXYZ function}
\details{An affine transformation of an object of class \code{fromXYZ}. Nearest Neighbour interpolation}
\value{the object acted on by the transformation matrix}
\arguments{\item{X}{an object of class fromFunction}
\item{mat}{matrix of affine transformation}
\item{...}{additional arguments}}
