\name{affine.stppp}
\alias{affine.stppp}
\title{affine...}
\usage{\method{affine}{stppp} (X, mat, ...)}
\description{affine.stppp function}
\details{An affine transformation of an object of class \code{stppp}}
\value{the object acted on by the transformation matrix}
\arguments{\item{X}{an object of class stppp}
\item{mat}{matrix of affine transformation}
\item{...}{additional arguments}}
