% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometryHandling.R
\name{aggCovInfo.ArealWeightedSum}
\alias{aggCovInfo.ArealWeightedSum}
\title{aggCovInfo.ArealWeightedSum function}
\usage{
\method{aggCovInfo}{ArealWeightedSum}(obj, regwts, ...)
}
\arguments{
\item{obj}{an ArealWeightedSum object}

\item{regwts}{regional (areal) weighting vector}

\item{...}{additional arguments}
}
\value{
Areal weighted Sum.
}
\description{
Aggregation via weighted sum. Use to sum up population counts in regions.
}
