\name{aggCovInfo.Majority}
\alias{aggCovInfo.Majority}
\title{aggCovInfo.Majority function}
\usage{
\method{aggCovInfo}{Majority}(obj, regwts, ...)
}
\arguments{
  \item{obj}{an Majority object}

  \item{regwts}{regional (areal) weighting vector}

  \item{...}{additional arguments}
}
\value{
The most popular cell type.
}
\description{
Aggregation via majority.
}

