% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometryHandling.R
\name{aggCovInfo}
\alias{aggCovInfo}
\title{aggCovInfo function}
\usage{
aggCovInfo(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{additional arguments}
}
\value{
method aggCovInfo
}
\description{
Generic function for aggregation of covariate information.
}
