% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometryHandling.R
\name{aggregateCovariateInfo}
\alias{aggregateCovariateInfo}
\title{aggregateCovariateInfo function}
\usage{
aggregateCovariateInfo(cellidx, cidx, gidx, df, fftovl, classes, polyareas)
}
\arguments{
\item{cellidx}{the index of the cell}

\item{cidx}{index of covariate, no longer used}

\item{gidx}{grid index}

\item{df}{the data frame containing the covariate information}

\item{fftovl}{an overlay of the fft grid onto the SpatialPolygonsDataFrame or SpatialPixelsDataFrame objects}

\item{classes}{vector of class attributes of the dataframe}

\item{polyareas}{polygon areas of the SpatialPolygonsDataFrame or SpatialPixelsDataFrame objects}
}
\value{
the interpolated covariate information onto the FFT grid
}
\description{
A function called by cov.interp.fft to allocate and perform interpolation of covariate infomation onto the FFT grid
}
