% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{aggregateformulaList}
\alias{aggregateformulaList}
\title{aggregateformulaList function}
\usage{
aggregateformulaList(x, ...)
}
\arguments{
\item{x}{an object of class "formulaList"}

\item{...}{other arguments}
}
\value{
a formula of the form X ~ var1 + var2 tec.
}
\description{
An internal function to collect terms from a formulalist. Not intended for general use.
}
