\name{as.array.lgcpgrid}
\alias{as.array.lgcpgrid}
\title{as...}
\usage{\method{as.array}{lgcpgrid} (x, ...)}
\description{as.array.lgcpgrid function}
\details{Method to convert an lgcpgrid object into an array.}
\value{conversion from lgcpgrid to array}
\arguments{\item{x}{an object of class lgcpgrid}
\item{...}{other arguments}}
