\name{as.im.fromXYZ}
\alias{as.im.fromXYZ}
\title{as...}
\usage{\method{as.im}{fromXYZ} (X, ...)}
\description{as.im.fromXYZ function}
\details{Convert an object of class fromXYZ (created by spatialAtRisk for example) into a spatstat im object.}
\value{object of class im containing normalised intensities}
\seealso{\link{as.im.fromSPDF}, \link{as.im.fromFunction}, \link{as.fromXYZ}}
\arguments{\item{X}{object of class fromXYZ}
\item{...}{additional arguments}}
