\name{as.list.lgcpgrid}
\alias{as.list.lgcpgrid}
\title{as...}
\usage{\method{as.list}{lgcpgrid} (x, ...)}
\description{as.list.lgcpgrid function}
\details{Method to convert an lgcpgrid object into a list of matrices.}
\value{conversion from lgcpgrid to list}
\seealso{\link{lgcpgrid.list}, \link{lgcpgrid.array}, \link{print.lgcpgrid}, 
\link{summary.lgcpgrid}, \link{quantile.lgcpgrid}, \link{image.lgcpgrid}, \link{plot.lgcpgrid}}
\arguments{\item{x}{an object of class lgcpgrid}
\item{...}{other arguments}}
