\name{as.owin.stapp}
\alias{as.owin.stapp}
\title{as...}
\usage{\method{as.owin}{stapp} (W, ..., fatal=TRUE)}
\description{as.owin.stapp function}
\details{A function to extract the SpatialPolygons part of W and return it as an owin object.}
\value{an owin object}
\arguments{\item{W}{see ?as.owin}
\item{...}{see ?as.owin}
\item{fatal}{see ?as.owin}}
