\name{as.ppp.mstppp}
\alias{as.ppp.mstppp}
\title{as...}
\usage{\method{as.ppp}{mstppp} (X, ..., fatal=TRUE)}
\description{as.ppp.mstppp function}
\details{Convert from mstppp to ppp. Can be useful for data handling.}
\value{a ppp object without observation times}
\arguments{\item{X}{an object of class mstppp}
\item{...}{additional arguments}
\item{fatal}{logical value, see details in generic ?as.ppp}}
