\name{as.ppp.stppp}
\alias{as.ppp.stppp}
\title{as...}
\usage{\method{as.ppp}{stppp} (X, ..., fatal=TRUE)}
\description{as.ppp.stppp function}
\details{Convert from stppp to ppp. Can be useful for data handling.}
\value{a ppp object without observation times}
\arguments{\item{X}{an object of class stppp}
\item{...}{additional arguments}
\item{fatal}{logical value, see details in generic ?as.ppp}}
