\name{as.SpatialPixelsDataFrame.lgcpgrid}
\alias{as.SpatialPixelsDataFrame.lgcpgrid}
\title{as...}
\usage{\method{as.SpatialPixelsDataFrame}{lgcpgrid} (obj, ...)}
\description{as.SpatialPixelsDataFrame.lgcpgrid function}
\details{Method to convert lgcpgrid objects to SpatialPixelsDataFrame objects.}
\value{Either a SpatialPixelsDataFrame, or a list consisting of SpatialPixelsDataFrame objects.}
\arguments{\item{obj}{an lgcpgrid object}
\item{...}{additional arguments to be passed to SpatialPoints, eg a proj4string}}
