\name{as.SpatialPixelsDataFrame}
\alias{as.SpatialPixelsDataFrame}
\title{as...}
\usage{as.SpatialPixelsDataFrame(obj, ...)}
\description{as.SpatialPixelsDataFrame function}
\details{Generic function for conversion to SpatialPixels objects.}
\value{method as.SpatialPixels}
\seealso{\link{as.SpatialPixelsDataFrame.lgcpgrid}}
\arguments{\item{obj}{an object}
\item{...}{additional arguments}}
