% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAtRiskClassDef.R
\name{as.fromXYZ.fromFunction}
\alias{as.fromXYZ.fromFunction}
\title{as.fromXYZ.fromFunction function}
\usage{
\method{as.fromXYZ}{fromFunction}(X, xyt, M = 100, N = 100, ...)
}
\arguments{
\item{X}{an object of class fromFunction}

\item{xyt}{and objects of class stppp}

\item{M}{number of cells in x direction}

\item{N}{number of cells in y direction}

\item{...}{additional arguments}
}
\value{
object of class im containing normalised intensities
}
\description{
Method for converting from the fromFunction class of objects to the fromXYZ class of objects. Clearly this requires the
user to specify a grid onto which to compute the discretised verion.
}
\seealso{
\link{as.im.fromXYZ}, \link{as.im.fromSPDF}, \link{as.im.fromFunction}, \link{as.fromXYZ}
}
