\name{as.fromXYZ}
\alias{as.fromXYZ}
\title{as...}
\usage{as.fromXYZ(X, ...)}
\description{as.fromXYZ function}
\details{Generic function for conversion to a fromXYZ object (eg as would have been produced by spatialAtRisk for example.)}
\value{generic function returning method as.fromXYZ}
\seealso{\link{as.im.fromXYZ}, \link{as.im.fromSPDF}, \link{as.im.fromFunction}, \link{as.fromXYZ}}
\arguments{\item{X}{an object}
\item{...}{additional arguments}}
