% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAtRiskClassDef.R
\name{as.im.fromFunction}
\alias{as.im.fromFunction}
\title{as.im.fromFunction function}
\usage{
\method{as.im}{fromFunction}(X, xyt, M = 100, N = 100, ...)
}
\arguments{
\item{X}{an object of class fromSPDF}

\item{xyt}{and objects of class stppp}

\item{M}{number of cells in x direction}

\item{N}{number of cells in y direction}

\item{...}{additional arguments}
}
\value{
object of class im containing normalised intensities
}
\description{
Convert an object of class fromFunction(created by spatialAtRisk for example) into a spatstat im object.
}
\seealso{
\link{as.im.fromXYZ}, \link{as.im.fromSPDF}, \link{as.im.fromFunction}, \link{as.fromXYZ}
}
