\name{as.im.fromXYZ}
\alias{as.im.fromXYZ}
\title{as.im.fromXYZ function}
\usage{
\method{as.im}{fromXYZ}(X, ...)
}
\arguments{
  \item{X}{object of class fromXYZ}

  \item{...}{additional arguments}
}
\value{
object of class im containing normalised intensities
}
\description{
Convert an object of class fromXYZ (created by
spatialAtRisk for example) into a spatstat im object.
}
\seealso{
\link{as.im.fromSPDF}, \link{as.im.fromFunction},
\link{as.fromXYZ}
}

