% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{as.list.lgcpgrid}
\alias{as.list.lgcpgrid}
\title{as.list.lgcpgrid function}
\usage{
\method{as.list}{lgcpgrid}(x, ...)
}
\arguments{
\item{x}{an object of class lgcpgrid}

\item{...}{other arguments}
}
\value{
conversion from lgcpgrid to list
}
\description{
Method to convert an lgcpgrid object into a list of matrices.
}
\seealso{
\link{lgcpgrid.list}, \link{lgcpgrid.array}, \link{print.lgcpgrid}, 
\link{summary.lgcpgrid}, \link{quantile.lgcpgrid}, \link{image.lgcpgrid}, \link{plot.lgcpgrid}
}
