% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arealMethods.R
\name{as.owinlist}
\alias{as.owinlist}
\title{as.owinlist function}
\usage{
as.owinlist(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{additional arguments}
}
\value{
method as.owinlist
}
\description{
Generic function for creating lists of owin objects
}
