% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstpppClassDef.R
\name{as.ppp.mstppp}
\alias{as.ppp.mstppp}
\title{as.ppp.mstppp function}
\usage{
\method{as.ppp}{mstppp}(X, ..., fatal = TRUE)
}
\arguments{
\item{X}{an object of class mstppp}

\item{...}{additional arguments}

\item{fatal}{logical value, see details in generic ?as.ppp}
}
\value{
a ppp object without observation times
}
\description{
Convert from mstppp to ppp. Can be useful for data handling.
}
