% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stpppClassDef.R
\name{as.ppp.stppp}
\alias{as.ppp.stppp}
\title{as.ppp.stppp function}
\usage{
\method{as.ppp}{stppp}(X, ..., fatal = TRUE)
}
\arguments{
\item{X}{an object of class stppp}

\item{...}{additional arguments}

\item{fatal}{logical value, see details in generic ?as.ppp}
}
\value{
a ppp object without observation times
}
\description{
Convert from stppp to ppp. Can be useful for data handling.
}
