\name{as.SpatialGridDataFrame.fromXYZ}
\alias{as.SpatialGridDataFrame.fromXYZ}
\title{as...}
\usage{\method{as.SpatialGridDataFrame}{fromXYZ} (obj, ...)}
\description{as.SpatialGridDataFrame.fromXYZ function}
\details{Method for converting objects of class fromXYZ into those of class SpatialGridDataFrame}
\value{an object of class SpatialGridDataFrame}
\seealso{\link{as.SpatialGridDataFrame}}
\arguments{\item{obj}{an object of class spatialAtRisk}
\item{...}{additional arguments}}
