\name{as.SpatialGridDataFrame}
\alias{as.SpatialGridDataFrame}
\title{as...}
\usage{as.SpatialGridDataFrame(obj, ...)}
\description{as.SpatialGridDataFrame function}
\details{Generic method for convertign to an object of class SpatialGridDataFrame.}
\value{method as.SpatialGridDataFrame}
\seealso{\link{as.SpatialGridDataFrame.fromXYZ}}
\arguments{\item{obj}{an object}
\item{...}{additional arguments}}
