% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{as.SpatialPixelsDataFrame}
\alias{as.SpatialPixelsDataFrame}
\title{as.SpatialPixelsDataFrame function}
\usage{
as.SpatialPixelsDataFrame(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{additional arguments}
}
\value{
method as.SpatialPixels
}
\description{
Generic function for conversion to SpatialPixels objects.
}
\seealso{
\link{as.SpatialPixelsDataFrame.lgcpgrid}
}
