% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arealMethods.R
\name{as.stppp}
\alias{as.stppp}
\title{as.stppp function}
\usage{
as.stppp(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{additional arguments}
}
\value{
method as.stppp
}
\description{
Generic function for converting to stppp objects
}
