% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpInternals.R
\name{at}
\alias{at}
\title{at function}
\usage{
at(t, mu, theta)
}
\arguments{
\item{t}{change in time parameter, see Brix and Diggle (2001)}

\item{mu}{mean}

\item{theta}{parameter beta in Brix and Diggle}
}
\value{
...
}
\description{
at function
}
