% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultitypeSpatialOnlyPlusParameters.R
\name{autocorrMultitype}
\alias{autocorrMultitype}
\title{autocorrMultitype function}
\usage{
autocorrMultitype(
  x,
  lags,
  fieldno,
  inWindow = x$xyt$window,
  crop2parentwindow = TRUE,
  ...
)
}
\arguments{
\item{x}{an object of class lgcpPredictMultitypeSpatialPlusParameters}

\item{lags}{the lags at which to compute the autocorrelation}

\item{fieldno}{the index of the lateyt field, the i in Y_i, see the help file for lgcpPredictMultitypeSpatialPlusParameters. IN diagnostic checking ,this command should be called for each field in the model.}

\item{inWindow}{an observation owin window on which to compute the autocorrelations, can speed up calculation. Default is x$xyt$window, set to NULL for full grid.}

\item{crop2parentwindow}{logical: whether to only compute autocorrelations for cells inside x$xyt$window (the 'parent window')}

\item{...}{other arguments}
}
\value{
an array, the [,,i]th slice being the grid of cell-wise autocorrelations.
}
\description{
A function to compute cell-wise autocorrelation in the latent field at specifiec lags
}
