% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{BetaParameters}
\alias{BetaParameters}
\title{BetaParameters function}
\usage{
BetaParameters(beta)
}
\arguments{
\item{beta}{a vector}
}
\value{
...
}
\description{
An internal function to declare a vector a parameter vector for the main effects.
}
