% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialOnly.R
\name{blockcircbase}
\alias{blockcircbase}
\title{blockcircbase function}
\usage{
blockcircbase(x, y, sigma, phi, model, additionalparameters, inverse = FALSE)
}
\arguments{
\item{x}{x centroids, an equally spaced vector}

\item{y}{y centroids, an equally spaced vector}

\item{sigma}{spatial variance parameter}

\item{phi}{spatial decay parameter}

\item{model}{covariance model, see ?CovarianceFct}

\item{additionalparameters}{additional parameters for chosen covariance model. See ?CovarianceFct}

\item{inverse}{logical. Whether to return the base matrix of the inverse covariance matrix (ie the base matrix for the precision matrix), default is FALSE}
}
\value{
the base matrix of a block circulant matrix representing a stationary covariance function on a toral grid.
}
\description{
Compute the base matrix of a continuous Gaussian field. Computed as a block circulant matrix on a torus where x and y is the 
x and y centroids (must be equally spaced)
}
