% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpInternals.R
\name{bt.scalar}
\alias{bt.scalar}
\title{bt.scalar function}
\usage{
bt.scalar(t, theta)
}
\arguments{
\item{t}{change in time, see Brix and Diggle (2001)}

\item{theta}{parameter beta in Brix and Diggle}
}
\value{
...
}
\description{
bt.scalar function
}
