% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{checkObsWin}
\alias{checkObsWin}
\title{checkObsWin function}
\usage{
checkObsWin(ow)
}
\arguments{
\item{ow}{an object generated by selectObsWindow, see ?selectObsWindow}
}
\value{
a plot of the observation window and grid
}
\description{
A function to run on an object generated by the "selectObsWindow" function. Plots the observation window with grid, use as a visual aid to check
the choice of cell width is correct.
}
\seealso{
\link{chooseCellwidth}
}
