\name{circulant.matrix}
\alias{circulant.matrix}
\title{circulant...}
\usage{\method{circulant}{matrix} (x, ...)}
\description{circulant.matrix function}
\details{If x is a matrix whose columns are the bases of the sub-blocks of a block circulant matrix, then this function returns the 
block circulant matrix of interest.}
\value{If x is a matrix whose columns are the bases of the sub-blocks of a block circulant matrix, then this function returns the block circulant matrix of interest.}
\arguments{\item{x}{a matrix object}
\item{...}{additional arguments}}
