\name{circulant.numeric}
\alias{circulant.numeric}
\title{circulant...}
\usage{\method{circulant}{numeric} (x, ...)}
\description{circulant.numeric function}
\details{returns a circulant matrix with base x}
\value{a circulant matrix with base x}
\arguments{\item{x}{an numeric object}
\item{...}{additional arguments}}
