% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometryHandling.R
\name{clearinterp}
\alias{clearinterp}
\title{clearinterp function}
\usage{
clearinterp(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
removes the interpolation methods
}
\description{
A function to remove the interpolation methods from a data frame.
}
