\name{constantInTime}
\alias{constantInTime}
\title{constantInTime function...}
\usage{constantInTime(obj, ...)}
\description{constantInTime function}
\details{Generic function for creating constant-in-time temporalAtRisk objects, that is for models where mu(t) can be assumed to
be constant in time. The assumption being that the global at-risk population does not change in size over time.

For further details of temporalAtRisk objects, see ?temporalAtRisk>}
\value{method constantInTime}
\seealso{\link{temporalAtRisk}, \link{spatialAtRisk}, \link{temporalAtRisk.numeric}, \link{temporalAtRisk.function}, \link{constantInTime.numeric}, \link{constantInTime.stppp}, \link{print.temporalAtRisk}, \link{plot.temporalAtRisk}}
\arguments{\item{obj}{an object}
\item{...}{additional arguments}}
