% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometryHandling.R
\name{cov.interp.fft}
\alias{cov.interp.fft}
\title{cov.interp.fft function}
\usage{
cov.interp.fft(
  formula,
  W,
  regionalcovariates = NULL,
  pixelcovariates = NULL,
  mcens,
  ncens,
  cellInside,
  overl = NULL
)
}
\arguments{
\item{formula}{an object of class formula (or one that can be coerced to that class) starting with X ~ (eg X~var1+var2 *NOT for example* Y~var1+var2): a symbolic description of the model to be fitted.}

\item{W}{an owin observation window}

\item{regionalcovariates}{an optional SpatialPolygonsDataFrame}

\item{pixelcovariates}{an optional SpatialPixelsDataFrame}

\item{mcens}{x-coordinates of output grid centroids (not fft grid centroids ie *not* the extended grid)}

\item{ncens}{y-coordinates of output grid centroids (not fft grid centroids ie *not* the extended grid)}

\item{cellInside}{a 0-1 matrix indicating which computational cells are inside the observation window}

\item{overl}{an overlay of the computational grid onto the SpatialPolygonsDataFrame or SpatialPixelsDataFrame.}
}
\value{
The interpolated design matrix, ready for analysis
}
\description{
A function to interpolate covariate values onto the fft grid, ready for analysis
}
