% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpInternals.R
\name{CovarianceFct}
\alias{CovarianceFct}
\title{CovarianceFct function}
\usage{
CovarianceFct(u, sigma, phi, model, additionalparameters)
}
\arguments{
\item{u}{distance}

\item{sigma}{parameter sigma}

\item{phi}{parameter phi}

\item{model}{character string, the model}

\item{additionalparameters}{additional parameters for the covariance function that will be fixed.}
}
\value{
the covariance function evaluated at the specified distances
}
\description{
A function to
}
