% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{CovFunction.function}
\alias{CovFunction.function}
\title{CovFunction.function function}
\usage{
\method{CovFunction}{`function`}(obj, ...)
}
\arguments{
\item{obj}{a function object}

\item{...}{additional arguments}
}
\value{
the covariance function ready to run the MCMC routine.
}
\description{
A function used to define the covariance function for the latent field prior to running the MCMC algorithm
}
\examples{
\dontrun{cf1 <- CovFunction(exponentialCovFct)}
\dontrun{cf2 <- CovFunction(RandomFieldsCovFct(model="matern",additionalparameters=1))}
}
\seealso{
\link{exponentialCovFct}, \link{RandomFieldsCovFct}, \link{SpikedExponentialCovFct}, \link{CovarianceFct}
}
