% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{CovFunction}
\alias{CovFunction}
\title{CovFunction function}
\usage{
CovFunction(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{additional arguments}
}
\value{
method CovFunction
}
\description{
A Generic method used to specify the choice of covariance function for use in the MCMC algorithm. For further details and examples,
see the vignette "Bayesian_lgcp".
}
\seealso{
\link{CovFunction.function}, \link{exponentialCovFct}, \link{RandomFieldsCovFct}, \link{SpikedExponentialCovFct}
}
