\name{density.stppp}
\alias{density.stppp}
\title{density...}
\usage{\method{density}{stppp} (x, bandwidth, ...)}
\description{density.stppp function}
\details{A wrapper function for \link{density.ppp}.}
\value{bivariate density estimate of xyt; not this is a wrapper function for density.ppp}
\seealso{\link{density.ppp}}
\arguments{\item{x}{an stppp object}
\item{bandwidth}{'bandwidth' parameter, equivanent to parameter sigma in ?density.ppp ie standard deviation of isotropic Gaussian smoothing kernel.}
\item{...}{additional arguments to be passed to density.ppp}}
