\name{discreteWindow.lgcpPredict}
\alias{discreteWindow.lgcpPredict}
\title{discreteWindow...}
\usage{\method{discreteWindow}{lgcpPredict} (obj, ...)}
\description{discreteWindow.lgcpPredict function}
\details{A function for extracting the FFT discrete window from an lgcpPredict object.}
\value{...}
\arguments{\item{obj}{an lgcpPredict object}
\item{...}{additional arguments}}
