\name{discreteWindow}
\alias{discreteWindow}
\title{discreteWindow function}
\usage{
  discreteWindow(obj, ...)
}
\arguments{
  \item{obj}{an object}

  \item{...}{additional arguments}
}
\value{
  method discreteWindow
}
\description{
  Generic function for extracting the FFT discrete window.
}
\seealso{
  \link{discreteWindow.lgcpPredict}
}

