\name{eigenfrombase}
\alias{eigenfrombase}
\title{eigenfrombase function...}
\usage{eigenfrombase(x)}
\description{eigenfrombase function}
\details{A function to compute the eigenvalues of an SPD block circulant matrix given the base matrix.}
\value{the eigenvalues}
\arguments{\item{x}{the base matrix}}
