% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{etavals}
\alias{etavals}
\title{etavals function}
\usage{
etavals(lg)
}
\arguments{
\item{lg}{an object produced by a call to lgcpPredictSpatialPlusPars, lgcpPredictAggregateSpatialPlusPars, lgcpPredictSpatioTemporalPlusPars orlgcpPredictMultitypeSpatialPlusPars}
}
\value{
the posterior sampled eta
}
\description{
A function to return the sampled eta from a call to the function lgcpPredictSpatialPlusPars, lgcpPredictAggregateSpatialPlusPars, lgcpPredictSpatioTemporalPlusPars or lgcpPredictMultitypeSpatialPlusPars
}
\seealso{
\link{ltar}, \link{autocorr}, \link{parautocorr}, \link{traceplots}, \link{parsummary}, \link{textsummary},
\link{priorpost}, \link{postcov}, \link{exceedProbs}, \link{betavals}
}
