\name{evallogprobY}
\alias{evallogprobY}
\title{evallogprobY function...}
\usage{evallogprobY(mat.list, at, bt, gmult, logdetcovY, Rinv)}
\description{evallogprobY function}
\value{...}
\arguments{\item{mat.list}{list of Y matrices}
\item{at}{(vector) component a_t from Brix and Diggle}
\item{bt}{(vector) component b_t from Brix and Diggle}
\item{gmult}{(vector) (1-exp(-2*beta*delta_t)) as in G(T); Brix and Diggle}
\item{logdetcovY}{log determinant of R matrix as in G(T); Brix and Diggle}
\item{Rinv}{inverse of R as in G(T); Brix and Diggle}}
