% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{EvaluatePrior}
\alias{EvaluatePrior}
\title{EvaluatePrior function}
\usage{
EvaluatePrior(etaParameters, betaParameters, prior)
}
\arguments{
\item{etaParameters}{the paramter eta}

\item{betaParameters}{the parameter beta}

\item{prior}{the prior}
}
\value{
the prior evaluated at the given values.
}
\description{
An internal function used in the MCMC routine to evaluate the prior for a given set of parameters
}
