% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{expectation.lgcpPredictSpatialOnlyPlusParameters}
\alias{expectation.lgcpPredictSpatialOnlyPlusParameters}
\title{expectation.lgcpPredictSpatialOnlyPlusParameters function}
\usage{
"expectation(obj,fun,maxit=NULL,...)"
}
\arguments{
\item{obj}{an object of class lgcpPredictSpatialOnlyPlusParameters}

\item{fun}{a function with arguments 'Y', 'beta', 'eta', 'Z' and 'otherargs'. See vignette("Bayesian_lgcp") for an example}

\item{maxit}{Not used in ordinary circumstances. Defines subset of samples over which to compute expectation. Expectation is computed using information from iterations 1:maxit, where 1 is the first non-burn in iteration dumped to disk.}

\item{...}{additional arguments}
}
\value{
the expectated value of that function
}
\description{
\bold{This function requires data to have been dumped to disk}: see \code{?dump2dir} and \code{?setoutput}. This function computes the 
Monte Carlo Average of a function where data from a run of \code{lgcpPredict} has been dumped to disk.
}
