\name{expectation}
\alias{expectation}
\title{expectation function...}
\usage{expectation(obj, ...)}
\description{expectation function}
\details{Generic function used in the computation of Monte Carlo expectations.}
\value{method expectation}
\arguments{\item{obj}{an object}
\item{...}{additional arguments}}
