\name{exponentialCovFct}
\alias{exponentialCovFct}
\title{exponentialCovFct function}
\usage{
exponentialCovFct(d, CovParameters)
}
\arguments{
  \item{d}{toral distance}

  \item{CovParameters}{parameters of the latent field, an
  object of class "CovParamaters".}
}
\value{
the exponential covariance function
}
\description{
A function to declare and also evaluate an exponential
covariance function.
}
\seealso{
\link{CovFunction.function}, \link{RandomFieldsCovFct},
\link{SpikedExponentialCovFct}
}

