\name{extract}
\alias{extract}
\title{extract function...}
\usage{extract(obj, ...)}
\description{extract function}
\details{Generic function for extracting information dumped to disk. See \link{extract.lgcpPredict} for further information.}
\value{method extract}
\seealso{\link{extract.lgcpPredict}}
\arguments{\item{obj}{an object}
\item{...}{additional arguments}}
