\name{Extract.stppp}
\title{Extract...}
\usage{"x[subset]"}
\description{Extract.stppp function}
\details{extracting subsets of an stppp object.}
\alias{"[.stppp"}
\value{extracts subset of an stppp object}
\arguments{\item{x}{an object of class stppp}
\item{...}{additional arguments}}
\examples{\dontrun{xyt <- lgcpSim()}
\dontrun{xyt}
\dontrun{xyt[xyt$t>0.5]}}
