\name{fftinterpolate.fromSPDF}
\alias{fftinterpolate.fromSPDF}
\title{fftinterpolate...}
\usage{\method{fftinterpolate}{fromSPDF} (spatial, mcens, ncens, ...)}
\description{fftinterpolate.fromSPDF function}
\details{This method performs interpolation within the function \code{fftgrid} for \code{fromSPDF} objects.}
\value{matrix of interpolated values}
\seealso{\link{fftgrid}, \link{spatialAtRisk.SpatialPolygonsDataFrame}}
\arguments{\item{spatial}{objects of class spatialAtRisk}
\item{mcens}{x-coordinates of interpolation grid in extended space}
\item{ncens}{y-coordinates of interpolation grid in extended space}
\item{...}{additional arguments}}
