\name{fftinterpolate}
\alias{fftinterpolate}
\title{fftinterpolate function...}
\usage{fftinterpolate(spatial, ...)}
\description{fftinterpolate function}
\details{Generic function used for computing interpolations used in the function \link{fftgrid}.}
\value{method fftinterpolate}
\seealso{\link{fftgrid}}
\arguments{\item{spatial}{an object}
\item{...}{additional arguments}}
