\name{fftmultiply}
\alias{fftmultiply}
\title{fftmultiply function}
\usage{
  fftmultiply(efb, vector)
}
\arguments{
  \item{efb}{eigenvalues of the matrix}

  \item{vector}{the vector}
}
\value{
  a vector: the product of the matrix and the vector.
}
\description{
  A function to pre-multiply a vector by a block cirulant
  matrix
}

