% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{formulaList}
\alias{formulaList}
\title{formulaList function}
\usage{
formulaList(X)
}
\arguments{
\item{X}{a list object, each element of which is a formula}
}
\value{
an object of class "formulaList"
}
\description{
A function to creat an object of class "formulaList" from a list of "formula" objects; use to define the model for the main effects
prior to running the multivariate MCMC algorithm.
}
