\name{GAfinalise.MonteCarloAverage}
\alias{GAfinalise.MonteCarloAverage}
\title{GAfinalise...}
\usage{\method{GAfinalise}{MonteCarloAverage} (F, ...)}
\description{GAfinalise.MonteCarloAverage function}
\details{Finalise a Monte Carlo averaging scheme. Divide the sum by the number of iterations.}
\value{computes Monte Carlo averages}
\seealso{\link{MonteCarloAverage}, \link{setoutput}, \link{GAinitialise}, \link{GAupdate}, \link{GAfinalise}, \link{GAreturnvalue}}
\arguments{\item{F}{an object of class MonteCarloAverage}
\item{...}{additional arguments}}
