\name{GAfinalise.nullAverage}
\alias{GAfinalise.nullAverage}
\title{GAfinalise...}
\usage{\method{GAfinalise}{nullAverage} (F, ...)}
\description{GAfinalise.nullAverage function}
\details{This is a null function and performs no action.}
\value{nothing}
\seealso{\link{nullAverage}, \link{setoutput}, \link{GAinitialise}, \link{GAupdate}, \link{GAfinalise}, \link{GAreturnvalue}}
\arguments{\item{F}{an object of class nullAverage}
\item{...}{additional arguments}}
