\name{GAfinalise}
\alias{GAfinalise}
\title{GAfinalise function...}
\usage{GAfinalise(F, ...)}
\description{GAfinalise function}
\details{Generic function defining the the finalisation step for the \code{gridAverage} class of functions.
The function is called invisibly within \code{MALAlgcp} and facilitates the computation of
Monte Carlo Averages online.}
\value{method GAfinalise}
\seealso{\link{setoutput}, \link{GAinitialise}, \link{GAupdate}, \link{GAreturnvalue}}
\arguments{\item{F}{an object}
\item{...}{additional arguments}}
