\name{GAinitialise.MonteCarloAverage}
\alias{GAinitialise.MonteCarloAverage}
\title{GAinitialise...}
\usage{\method{GAinitialise}{MonteCarloAverage} (F, ...)}
\description{GAinitialise.MonteCarloAverage function}
\details{Initialise a Monte Carlo averaging scheme.}
\value{nothing}
\seealso{\link{MonteCarloAverage}, \link{setoutput}, \link{GAinitialise}, \link{GAupdate}, \link{GAfinalise}, \link{GAreturnvalue}}
\arguments{\item{F}{an object of class MonteCarloAverage}
\item{...}{additional arguments}}
