\name{GAinitialise.MonteCarloAverage}
\alias{GAinitialise.MonteCarloAverage}
\title{GAinitialise.MonteCarloAverage function}
\usage{
  \method{GAinitialise}{MonteCarloAverage} (F, ...)
}
\arguments{
  \item{F}{an object of class MonteCarloAverage}

  \item{...}{additional arguments}
}
\value{
  nothing
}
\description{
  Initialise a Monte Carlo averaging scheme.
}
\seealso{
  \link{MonteCarloAverage}, \link{setoutput},
  \link{GAinitialise}, \link{GAupdate}, \link{GAfinalise},
  \link{GAreturnvalue}
}

