\name{GAinitialise}
\alias{GAinitialise}
\title{GAinitialise function...}
\usage{GAinitialise(F, ...)}
\description{GAinitialise function}
\details{Generic function defining the the initialisation step for the \code{gridAverage} class of functions. 
The function is called invisibly within \code{MALAlgcp} and facilitates the computation of
Monte Carlo Averages online.}
\value{method GAinitialise}
\seealso{\link{setoutput}, \link{GAupdate}, \link{GAfinalise}, \link{GAreturnvalue}}
\arguments{\item{F}{an object}
\item{...}{additional arguments}}
