\name{GammafromY}
\alias{GammafromY}
\title{GammafromY function...}
\usage{GammafromY(Y, rootQeigs, mu)}
\description{GammafromY function}
\details{A function to change Ys (spatially correlated noise) into Gammas (white noise). Used in the MALA algorithm.}
\value{Gamma}
\arguments{\item{Y}{Y matrix}
\item{rootQeigs}{square root of the eigenvectors of the precision matrix}
\item{mu}{parameter of the latent Gaussian field}}
