\name{GAreturnvalue.nullAverage}
\alias{GAreturnvalue.nullAverage}
\title{GAreturnvalue...}
\usage{\method{GAreturnvalue}{nullAverage} (F, ...)}
\description{GAreturnvalue.nullAverage function##'}
\details{This is a null function and performs no action.}
\value{nothing}
\seealso{\link{nullAverage}, \link{setoutput}, \link{GAinitialise}, \link{GAupdate}, \link{GAfinalise}, \link{GAreturnvalue}}
\arguments{\item{F}{an object of class nullAverage}
\item{...}{additional arguments}}
