\name{GAreturnvalue.MonteCarloAverage}
\alias{GAreturnvalue.MonteCarloAverage}
\title{GAreturnvalue...}
\usage{\method{GAreturnvalue}{MonteCarloAverage} (F, ...)}
\description{GAreturnvalue.MonteCarloAverage function}
\details{Returns the required Monte Carlo average.}
\value{results from MonteCarloAverage}
\seealso{\link{MonteCarloAverage}, \link{setoutput}, \link{GAinitialise}, \link{GAupdate}, \link{GAfinalise}, \link{GAreturnvalue}}
\arguments{\item{F}{an object of class MonteCarloAverage}
\item{...}{additional arguments}}
