\name{GAreturnvalue}
\alias{GAreturnvalue}
\title{GAreturnvalue function...}
\usage{GAreturnvalue(F, ...)}
\description{GAreturnvalue function}
\details{Generic function defining the the returned value for the \code{gridAverage} class of functions.
The function is called invisibly within \code{MALAlgcp} and facilitates the computation of
Monte Carlo Averages online.}
\value{method GAreturnvalue}
\seealso{\link{setoutput}, \link{GAinitialise}, \link{GAupdate}, \link{GAfinalise}}
\arguments{\item{F}{an object}
\item{...}{additional arguments}}
