\name{GAupdate.MonteCarloAverage}
\alias{GAupdate.MonteCarloAverage}
\title{GAupdate...}
\usage{\method{GAupdate}{MonteCarloAverage} (F, ...)}
\description{GAupdate.MonteCarloAverage function}
\details{Update a Monte Carlo averaging scheme. This function performs the Monte Carlo sum online.}
\value{updates Monte Carlo sums}
\seealso{\link{MonteCarloAverage}, \link{setoutput}, \link{GAinitialise}, \link{GAupdate}, \link{GAfinalise}, \link{GAreturnvalue}}
\arguments{\item{F}{an object of class MonteCarloAverage}
\item{...}{additional arguments}}
