\name{GAupdate.nullAverage}
\alias{GAupdate.nullAverage}
\title{GAupdate...}
\usage{\method{GAupdate}{nullAverage} (F, ...)}
\description{GAupdate.nullAverage function}
\details{This is a null function and performs no action.}
\value{nothing}
\seealso{\link{nullAverage}, \link{setoutput}, \link{GAinitialise}, \link{GAupdate}, \link{GAfinalise}, \link{GAreturnvalue}}
\arguments{\item{F}{an object of class nullAverage}
\item{...}{additional arguments}}
