% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridAverageMethods.R
\name{GAupdate.MonteCarloAverage}
\alias{GAupdate.MonteCarloAverage}
\title{GAupdate.MonteCarloAverage function}
\usage{
\method{GAupdate}{MonteCarloAverage}(F, ...)
}
\arguments{
\item{F}{an object of class MonteCarloAverage}

\item{...}{additional arguments}
}
\value{
updates Monte Carlo sums
}
\description{
Update a Monte Carlo averaging scheme. This function performs the Monte Carlo sum online.
}
\seealso{
\link{MonteCarloAverage}, \link{setoutput}, \link{GAinitialise}, \link{GAupdate}, \link{GAfinalise}, \link{GAreturnvalue}
}
