\name{GAupdate}
\alias{GAupdate}
\title{GAupdate function...}
\usage{GAupdate(F, ...)}
\description{GAupdate function}
\details{Generic function defining the the update step for the \code{gridAverage} class of functions.
The function is called invisibly within \code{MALAlgcp} and facilitates the computation of
Monte Carlo Averages online.}
\value{method GAupdate}
\seealso{\link{setoutput}, \link{GAinitialise}, \link{GAfinalise}, \link{GAreturnvalue}}
\arguments{\item{F}{an object}
\item{...}{additional arguments}}
