% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{GaussianPrior}
\alias{GaussianPrior}
\title{GaussianPrior function}
\usage{
GaussianPrior(mean, variance)
}
\arguments{
\item{mean}{a vector of length 2 representing the mean.}

\item{variance}{a 2x2 matrix representing the variance.}
}
\value{
an object of class LogGaussianPrior that can be passed to the function PriorSpec.
}
\description{
A function to create a Gaussian prior.
}
\examples{
\dontrun{GaussianPrior(mean=rep(0,9),variance=diag(10^6,9))}
}
\seealso{
\link{LogGaussianPrior}, link{PriorSpec.list}
}
