% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{genFFTgrid}
\alias{genFFTgrid}
\title{genFFTgrid function}
\usage{
genFFTgrid(study.region, M, N, ext, inclusion = "touching")
}
\arguments{
\item{study.region}{an owin object}

\item{M}{number of cells in x direction}

\item{N}{number of cells in y direction}

\item{ext}{multiplying constant: the size of the extended grid: ext*M by ext*N}

\item{inclusion}{criterion for cells being included into observation window. Either 'touching' or 'centroid'. The former includes all cells that touch the observation window, the latter includes all cells whose centroids are inside the observation window.}
}
\value{
a list
}
\description{
A function to generate an FFT grid and associated quantities including cell dimensions,
size of extended grid, centroids, cell area, cellInside matrix (a 0/1 matrix: is the centroid of the cell inside the observation window?)
}
