\name{getCellCounts}
\alias{getCellCounts}
\title{getCellCounts function...}
\usage{getCellCounts(x, y, xgrid, ygrid)}
\description{getCellCounts function}
\details{This function is used to count the number of observations falling inside grid cells.}
\value{The number of observations in each grid cell.}
\arguments{\item{x}{x-coordinates of events}
\item{y}{y-coordinates of events}
\item{xgrid}{x-coordinates of grid centroids}
\item{ygrid}{y-coordinates of grid centroids}}
