\name{getCovParameters.GPrealisation}
\alias{getCovParameters.GPrealisation}
\title{getCovParameters.GPrealisation function}
\usage{
  \method{getCovParameters}{GPrealisation} (obj, ...)
}
\arguments{
  \item{obj}{an GPrealisation object}

  \item{...}{additional arguments}
}
\value{
  ...
}
\description{
  Internal function for retrieving covariance parameters.
  not indended for general use.
}

