% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometryHandling.R
\name{getinterp}
\alias{getinterp}
\title{getinterp function}
\usage{
getinterp(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
the interpolation methods
}
\description{
A function to get the interpolation methods from a data frame\cr
}
\details{
The three types of interpolation method employed in the package lgcp are:\cr
\enumerate{
   \item 'Majority' The interpolated value corresponds to the value of the covariate occupying
       the largest area of the computational cell.
   \item 'ArealWeightedMean' The interpolated value corresponds to the mean of all covariate
       values contributing to the computational cell weighted by their respective areas.
   \item 'ArealWeightedSum' The interpolated value is the sum of all contributing covariates
       weighed by the proportion of area with respect to the covariate polygons. For example,
       suppose region A has the same area as a computational grid cell and has 500 inhabitants.
       If that region occupies half of a computational grid cell, then this interpolation type assigns
       250 inhabitants from A to the computational grid cell.
}
}
