% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lgcp-INLA.R
\name{getlgcpPredictSpatialINLA}
\alias{getlgcpPredictSpatialINLA}
\title{getlgcpPredictSpatialINLA function}
\usage{
getlgcpPredictSpatialINLA()
}
\value{
Does not return anything
}
\description{
A function to download and 'install' lgcpPredictSpatialINLA into the lgcp namespace.
}

