% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{getLHSformulaList}
\alias{getLHSformulaList}
\title{getLHSformulaList function}
\usage{
getLHSformulaList(fl)
}
\arguments{
\item{fl}{an object of class "formulaList"}
}
\value{
the indepentdent variables
}
\description{
A function to retrieve the dependent variables from a formulaList object. Not intended for general use.
}
