\name{getRotation.default}
\alias{getRotation.default}
\title{getRotation...}
\usage{\method{getRotation}{default} (xyt, ...)}
\description{getRotation.default function}
\details{Presently there is no default method, see ?getRotation.stppp}
\value{currently no default implementation}
\seealso{\link{getRotation.stppp}}
\arguments{\item{xyt}{an object}
\item{...}{additional arguments}}
