% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectObsWindow.R
\name{getRotation.default}
\alias{getRotation.default}
\title{getRotation.default function}
\usage{
\method{getRotation}{default}(xyt, ...)
}
\arguments{
\item{xyt}{an object}

\item{...}{additional arguments}
}
\value{
currently no default implementation
}
\description{
Presently there is no default method, see ?getRotation.stppp
}
\seealso{
\link{getRotation.stppp}
}
