\name{getRotation}
\alias{getRotation}
\title{getRotation function...}
\usage{getRotation(xyt, ...)}
\description{getRotation function}
\details{Generic function for the computation of rotation matrices.}
\value{method getRotation}
\seealso{\link{getRotation.stppp}}
\arguments{\item{xyt}{an object}
\item{...}{additional arguments}}
