% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{getup}
\alias{getup}
\title{getup function}
\usage{
getup(n, lev = 1)
}
\arguments{
\item{n}{a character string, the name of the object}

\item{lev}{how many levels up the hierarchy to go (see the argument "envir" from the function "get"), default is 1.}
}
\value{
...
}
\description{
A function to get an object from a parent frame.
}
