% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{getZmats}
\alias{getZmats}
\title{getZmats function}
\usage{
getZmats(Zmat, formulaList)
}
\arguments{
\item{Zmat}{an objecty of class "lgcpZmat"}

\item{formulaList}{an object of class "formulaList"}
}
\value{
design matrices for each of the point types
}
\description{
An internal function to create Z_k from an lgcpZmat object, for use in the multivariate MCMC algorithm. Not intended for general use.
}
