\name{GFfinalise.dump2dir}
\alias{GFfinalise.dump2dir}
\title{GFfinalise...}
\usage{\method{GFfinalise}{dump2dir} (F, ...)}
\description{GFfinalise.dump2dir function}
\details{This function finalises the dumping of data to a netCDF file.}
\value{nothing}
\seealso{\link{dump2dir}, \link{setoutput}, \link{GFinitialise}, \link{GFupdate}, \link{GFfinalise}, \link{GFreturnvalue}}
\arguments{\item{F}{an object}
\item{...}{additional arguments}}
