\name{GFfinalise.nullFunction}
\alias{GFfinalise.nullFunction}
\title{GFfinalise...}
\usage{\method{GFfinalise}{nullFunction} (F, ...)}
\description{GFfinalise.nullFunction function}
\details{This is a null function and performs no action.}
\value{nothing}
\seealso{\link{nullFunction}, \link{setoutput}, \link{GFinitialise}, \link{GFupdate}, \link{GFfinalise}, \link{GFreturnvalue}}
\arguments{\item{F}{an object of class dump2dir}
\item{...}{additional arguments}}
