% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridFunctionMethods.R
\name{GFfinalise.dump2dir}
\alias{GFfinalise.dump2dir}
\title{GFfinalise.dump2dir function}
\usage{
\method{GFfinalise}{dump2dir}(F, ...)
}
\arguments{
\item{F}{an object}

\item{...}{additional arguments}
}
\value{
nothing
}
\description{
This function finalises the dumping of data to a netCDF file.
}
\seealso{
\link{dump2dir}, \link{setoutput}, \link{GFinitialise}, \link{GFupdate}, \link{GFfinalise}, \link{GFreturnvalue}
}
