\name{GFfinalise}
\alias{GFfinalise}
\title{GFfinalise function...}
\usage{GFfinalise(F, ...)}
\description{GFfinalise function}
\details{Generic function defining the the finalisation step for the \code{gridFunction} class of objects. 
The function is called invisibly within \code{MALAlgcp} and facilitates the dumping of data to disk}
\value{method GFfinalise}
\seealso{\link{setoutput}, \link{GFinitialise}, \link{GFupdate}, \link{GFreturnvalue}}
\arguments{\item{F}{an object}
\item{...}{additional arguments}}
