\name{GFinitialise.dump2dir}
\alias{GFinitialise.dump2dir}
\title{GFinitialise...}
\usage{\method{GFinitialise}{dump2dir} (F, ...)}
\description{GFinitialise.dump2dir function}
\details{Creates a directory (if necessary) and allocates space for a netCDF dump.}
\value{creates initialisation file and folder}
\seealso{\link{dump2dir}, \link{setoutput}, \link{GFinitialise}, \link{GFupdate}, \link{GFfinalise}, \link{GFreturnvalue}}
\arguments{\item{F}{an object of class dump2dir}
\item{...}{additional arguments}}
