\name{GFinitialise.nullFunction}
\alias{GFinitialise.nullFunction}
\title{GFinitialise...}
\usage{\method{GFinitialise}{nullFunction} (F, ...)}
\description{GFinitialise.nullFunction function}
\details{This is a null function and performs no action.}
\value{nothing}
\seealso{\link{nullFunction}, \link{setoutput}, \link{GFinitialise}, \link{GFupdate}, \link{GFfinalise}, \link{GFreturnvalue}}
\arguments{\item{F}{an object of class dump2dir}
\item{...}{additional arguments}}
