\name{GFreturnvalue.dump2dir}
\alias{GFreturnvalue.dump2dir}
\title{GFreturnvalue...}
\usage{\method{GFreturnvalue}{dump2dir} (F, ...)}
\description{GFreturnvalue.dump2dir function}
\details{This function returns the name of the directory the netCDF file was written to.}
\value{display where files have been written to}
\seealso{\link{dump2dir}, \link{setoutput}, \link{GFinitialise}, \link{GFupdate}, \link{GFfinalise}, \link{GFreturnvalue}}
\arguments{\item{F}{an object}
\item{...}{additional arguments}}
