% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridFunctionMethods.R
\name{GFreturnvalue.nullFunction}
\alias{GFreturnvalue.nullFunction}
\title{GFreturnvalue.nullFunction function}
\usage{
\method{GFreturnvalue}{nullFunction}(F, ...)
}
\arguments{
\item{F}{an object of class dump2dir}

\item{...}{additional arguments}
}
\value{
nothing
}
\description{
This is a null function and performs no action.
}
\seealso{
\link{nullFunction}, \link{setoutput}, \link{GFinitialise}, \link{GFupdate}, \link{GFfinalise}, \link{GFreturnvalue}
}
