\name{GFreturnvalue}
\alias{GFreturnvalue}
\title{GFreturnvalue function...}
\usage{GFreturnvalue(F, ...)}
\description{GFreturnvalue function}
\details{Generic function defining the the returned value for the \code{gridFunction} class of objects. 
The function is called invisibly within \code{MALAlgcp} and facilitates the dumping of data to disk}
\value{method GFreturnvalue}
\seealso{\link{setoutput}, \link{GFinitialise}, \link{GFupdate}, \link{GFfinalise}}
\arguments{\item{F}{an object}
\item{...}{additional arguments}}
