\name{GFupdate.dump2dir}
\alias{GFupdate.dump2dir}
\title{GFupdate...}
\usage{\method{GFupdate}{dump2dir} (F, ...)}
\description{GFupdate.dump2dir function}
\details{This function gets the required information from \code{MALAlgcp} and writes the data to the netCDF file.}
\value{saves latent field}
\seealso{\link{dump2dir}, \link{setoutput}, \link{GFinitialise}, \link{GFupdate}, \link{GFfinalise}, \link{GFreturnvalue}}
\arguments{\item{F}{an object}
\item{...}{additional arguments}}
