\name{GFupdate}
\alias{GFupdate}
\title{GFupdate function...}
\usage{GFupdate(F, ...)}
\description{GFupdate function}
\details{Generic function defining the the update step for the \code{gridFunction} class of objects. 
The function is called invisibly within \code{MALAlgcp} and facilitates the dumping of data to disk}
\value{method GFupdate}
\seealso{\link{setoutput}, \link{GFinitialise}, \link{GFfinalise}, \link{GFreturnvalue}}
\arguments{\item{F}{an object}
\item{...}{additional arguments}}
