% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{GPdrv2_Multitype}
\alias{GPdrv2_Multitype}
\title{GPdrv2_Multitype function}
\usage{
GPdrv2_Multitype(
  GPlist,
  priorlist,
  Zlist,
  Ztlist,
  etalist,
  betalist,
  nis,
  cellarea,
  spatial,
  gradtrunc,
  fftgrid,
  covfunction,
  d,
  eps = 1e-06,
  k
)
}
\arguments{
\item{GPlist}{a list of objects of class GPrealisation}

\item{priorlist}{list of priors for the model}

\item{Zlist}{list of design matirices on the FFT grid}

\item{Ztlist}{list of transpose design matrices}

\item{etalist}{list of parameters, eta, for each realisation}

\item{betalist}{clist of parameters, beta, for each realisation}

\item{nis}{cell counts of each type the extended grid}

\item{cellarea}{the cell area}

\item{spatial}{list of poisson offsets for each type}

\item{gradtrunc}{gradient truncation parameter}

\item{fftgrid}{an object of class FFTgrid}

\item{covfunction}{list giving the choice of covariance function for each type, see ?CovFunction}

\item{d}{matrix of toral distances}

\item{eps}{the finite difference step size}

\item{k}{index of type for which to compute the gradient and hessian}
}
\value{
first and second derivatives of the log target  for tyupe k at the specified paramters Y, eta and beta
}
\description{
A function to compute the second derivatives of the log target for the multivariate model with respect to the paramters of the latent field. Not intended for general use.
}
