% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{GPlist2array}
\alias{GPlist2array}
\title{GPlist2array function}
\usage{
GPlist2array(GPlist, element)
}
\arguments{
\item{GPlist}{an object of class GPrealisation}

\item{element}{the name of the element of GPlist[[1]] (for example) to extract, e.g. "Y"}
}
\value{
an array
}
\description{
An internal function for turning a list of GPrealisation objects into an an array by a particular common element of the GPrealisation object
}
