% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{GPrealisation}
\alias{GPrealisation}
\title{GPrealisation function}
\usage{
GPrealisation(gamma, fftgrid, covFunction, covParameters, d)
}
\arguments{
\item{gamma}{the transformed (white noise) realisation of the process}

\item{fftgrid}{an object of class FFTgrid, see ?genFFTgrid}

\item{covFunction}{an object of class function returning the spatial covariance}

\item{covParameters}{an object of class CovParamaters, see ?CovParamaters}

\item{d}{matrix of grid distances}
}
\value{
a realisation of a spatial Gaussian process on a regular grid
}
\description{
A function to store a realisation of a spatial gaussian process for use in MCMC algorithms that include Bayesian parameter estimation.
Stores not only the realisation, but also computational quantities.
}
