\name{grad.comp}
\alias{grad.comp}
\title{grad...}
\usage{grad.comp(m, i, theta, varvals, nicounts.mat.list, timediffs, car.mat,
    ys, spatial, temporal)}
\description{grad.comp function}
\value{gradient}
\arguments{\item{m}{parameter}
\item{i}{index}
\item{theta}{beta parameter in Brix and Diggle (2001)}
\item{varvals}{pre computed variances}
\item{nicounts.mat.list}{matrix of nis}
\item{timediffs}{vector of delta ts}
\item{car.mat}{matrix}
\item{ys}{Gaussian field}
\item{spatial}{fixed spatial component}
\item{temporal}{fixed temporal component}}
