\name{gridav.lgcpPredict}
\alias{gridav.lgcpPredict}
\title{gridav...}
\usage{\method{gridav}{lgcpPredict} (obj, fun, ...)}
\description{gridav.lgcpPredict function}
\details{Accessor function for \code{lgcpPredict objects}: returns the \code{gridmeans} argument
set in the \code{output.control} argument of the function \code{lgcpPredict}.}
\value{returns the output from the gridmeans option of the setoutput argument of lgcpPredict}
\seealso{\link{setoutput}, \link{lgcpgrid}}
\arguments{\item{obj}{an object of class lgcpPredict}
\item{fun}{an optional character vector of length 1 giving the name of a function to return Monte Carlo average of}
\item{...}{additional arguments}}
