% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{gridav.lgcpPredict}
\alias{gridav.lgcpPredict}
\title{gridav.lgcpPredict function}
\usage{
\method{gridav}{lgcpPredict}(obj, fun = NULL, ...)
}
\arguments{
\item{obj}{an object of class lgcpPredict}

\item{fun}{an optional character vector of length 1 giving the name of a function to return Monte Carlo average of}

\item{...}{additional arguments}
}
\value{
returns the output from the gridmeans option of the setoutput argument of lgcpPredict
}
\description{
Accessor function for \code{lgcpPredict objects}: returns the \code{gridmeans} argument
set in the \code{output.control} argument of the function \code{lgcpPredict}.
}
\seealso{
\link{setoutput}, \link{lgcpgrid}
}
