\name{gridav}
\alias{gridav}
\title{gridav function...}
\usage{gridav(obj, ...)}
\description{gridav function}
\details{A generic function for returning \code{gridmeans} objects.}
\value{method gridav}
\seealso{\link{setoutput}, \link{lgcpgrid}}
\arguments{\item{obj}{an object}
\item{...}{additional arguments}}
