\name{gridfun.lgcpPredict}
\alias{gridfun.lgcpPredict}
\title{gridfun...}
\usage{\method{gridfun}{lgcpPredict} (obj, ...)}
\description{gridfun.lgcpPredict function}
\details{Accessor function for \code{lgcpPredict objects}: returns the \code{gridfunction} argument
set in the \code{output.control} argument of the function \code{lgcpPredict}.}
\value{returns the output from the gridfunction option of the setoutput argument of lgcpPredict}
\seealso{\link{setoutput}, \link{lgcpgrid}}
\arguments{\item{obj}{an object of class lgcpPredict}
\item{...}{additional arguments}}
