% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{gridfun.lgcpPredict}
\alias{gridfun.lgcpPredict}
\title{gridfun.lgcpPredict function}
\usage{
\method{gridfun}{lgcpPredict}(obj, ...)
}
\arguments{
\item{obj}{an object of class lgcpPredict}

\item{...}{additional arguments}
}
\value{
returns the output from the gridfunction option of the setoutput argument of lgcpPredict
}
\description{
Accessor function for \code{lgcpPredict objects}: returns the \code{gridfunction} argument
set in the \code{output.control} argument of the function \code{lgcpPredict}.
}
\seealso{
\link{setoutput}, \link{lgcpgrid}
}
