\name{gridfun}
\alias{gridfun}
\title{gridfun function...}
\usage{gridfun(obj, ...)}
\description{gridfun function}
\details{A generic function for returning \code{gridfunction} objects.}
\value{method gridfun}
\seealso{\link{setoutput}, \link{lgcpgrid}}
\arguments{\item{obj}{an object}
\item{...}{additional arguments}}
