\name{gu}
\alias{gu}
\title{gu function...}
\usage{gu(u, sigma, phi, model, additionalparameters)}
\description{gu function}
\value{this is just a wrapper for CovarianceFct}
\arguments{\item{u}{distance}
\item{sigma}{variance parameter, see Brix and Diggle (2001)}
\item{phi}{scale parameter, see Brix and Diggle (2001)}
\item{model}{correlation type, see ?CovarianceFct}
\item{additionalparameters}{vector of additional parameters for certain classes of covariance function (eg Matern), these must be supplied in the order given in ?CovarianceFct}}
