% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcIterator.R
\name{hasNext.iter}
\alias{hasNext.iter}
\title{hasNext.iter function}
\usage{
\method{hasNext}{iter}(obj)
}
\arguments{
\item{obj}{an iterator}
}
\description{
method for iter objects
test if an iterator has any more values to go
}
