\name{hvals.lgcpPredict}
\alias{hvals.lgcpPredict}
\title{hvals...}
\usage{\method{hvals}{lgcpPredict} (obj, ...)}
\description{hvals.lgcpPredict function}
\details{Accessor function returning the value of \eqn{h}{h}, the MALA proposal scaling constant over the iterations of the algorithm for 
objects of class \code{lgcpPredict}}
\value{returns the values of h taken during the progress of the algorithm}
\seealso{\link{lgcpPredict}}
\arguments{\item{obj}{an object of class lgcpPredict}
\item{...}{additional arguments}}
