\name{hvals}
\alias{hvals}
\title{hvals function...}
\usage{hvals(obj, ...)}
\description{hvals function}
\details{Generic function to return the values of the proposal scaling \eqn{h}{h} in the MCMC algorithm.}
\value{method hvals}
\arguments{\item{obj}{an object}
\item{...}{additional arguments}}
