\name{identifygrid}
\alias{identifygrid}
\title{identifygrid function...}
\usage{identifygrid(x, y)}
\description{identifygrid function}
\details{Identifies the indices of grid cells on plots of objects.}
\value{a 2 x n matrix containing the grid indices of the points of interest, where n is the number of points selected via the mouse.}
\seealso{\link{lgcpPredict}, \link{loc2poly}, \link{identify.lgcpPredict}}
\arguments{\item{x}{the x grid centroids}
\item{y}{the y grid centroids}}
