\name{initialiseAMCMC.constanth}
\alias{initialiseAMCMC.constanth}
\title{initaliseAMCMC...}
\usage{\method{initialiseAMCMC}{constanth} (obj, ...)}
\description{initaliseAMCMC.constanth function}
\details{Initialises the \link{constanth} adaptive scheme.}
\value{initial h for scheme}
\seealso{\link{constanth}}
\arguments{\item{obj}{an object}
\item{...}{additional arguments}}
