\name{initialiseAMCMC}
\alias{initialiseAMCMC}
\title{initialiseAMCMC function...}
\usage{initialiseAMCMC(obj, ...)}
\description{initialiseAMCMC function}
\details{A generic to be used for the purpose of user-defined adaptive MCMC schemes,
initialiseAMCMC tells the MALA algorithm which value of h to use first. See
lgcp vignette, code{vignette("lgcp")}, for further details on writing adaptive MCMC schemes.}
\value{method intialiseAMCMC}
\seealso{\link{initialiseAMCMC.constanth}, \link{initialiseAMCMC.andrieuthomsh}}
\arguments{\item{obj}{an object}
\item{...}{additional arguments}}
