\name{integerise.mstppp}
\alias{integerise.mstppp}
\title{integerise...}
\usage{\method{integerise}{mstppp} (obj, ...)}
\description{integerise.mstppp function}
\details{Function for converting the times and time limits of an mstppp object into integer values.}
\value{The mstppp object, but with integerised times.}
\arguments{\item{obj}{an mstppp object}
\item{...}{additional arguments}}
