\name{integerise.mstppp}
\alias{integerise.mstppp}
\title{integerise.mstppp function}
\usage{
\method{integerise}{mstppp}(obj, ...)
}
\arguments{
  \item{obj}{an mstppp object}

  \item{...}{additional arguments}
}
\value{
The mstppp object, but with integerised times.
}
\description{
Function for converting the times and time limits of an
mstppp object into integer values.
}

