\name{integerise}
\alias{integerise}
\title{integerise function...}
\usage{integerise(obj, ...)}
\description{integerise function}
\details{Generic function for converting the time variable of an stppp object.}
\value{method integerise}
\seealso{\link{integerise.stppp}}
\arguments{\item{obj}{an object}
\item{...}{additional arguments}}
