\name{intens.lgcpPredict}
\alias{intens.lgcpPredict}
\title{intens...}
\usage{\method{intens}{lgcpPredict} (obj, ...)}
\description{intens.lgcpPredict function}
\details{Accessor function returning the Poisson intensity as an lgcpgrid object.}
\value{the cell-wise mean Poisson intensity, as computed by MCMC.}
\seealso{\link{lgcpPredict}}
\arguments{\item{obj}{an lgcpPredict object}
\item{...}{additional arguments}}
