% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{intens.lgcpPredict}
\alias{intens.lgcpPredict}
\title{intens.lgcpPredict function}
\usage{
\method{intens}{lgcpPredict}(obj, ...)
}
\arguments{
\item{obj}{an lgcpPredict object}

\item{...}{additional arguments}
}
\value{
the cell-wise mean Poisson intensity, as computed by MCMC.
}
\description{
Accessor function returning the Poisson intensity as an lgcpgrid object.
}
\seealso{
\link{lgcpPredict}
}
