% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultitypeSpatialOnlyPlusParameters.R
\name{intens.lgcpSimMultitypeSpatialPlusParameters}
\alias{intens.lgcpSimMultitypeSpatialPlusParameters}
\title{intens.lgcpSimMultitypeSpatialPlusParameters function}
\usage{
"intens(obj, ...)"
}
\arguments{
\item{obj}{an object of class lgcpSimMultitypeSpatialPlusParameters}

\item{...}{other parameters}
}
\value{
the Poisson intensity
}
\description{
A function to return the cellwise Poisson intensity used during in constructing the simulated data.
}
