\name{intens}
\alias{intens}
\title{intens function}
\usage{
intens(obj, ...)
}
\arguments{
  \item{obj}{an object}

  \item{...}{additional arguments}
}
\value{
method intens
}
\description{
Generic function to return the Poisson Intensity.
}
\seealso{
\link{lgcpPredict}, \link{intens.lgcpPredict}
}

