\name{inversebase}
\alias{inversebase}
\title{inversebase function...}
\usage{inversebase(x)}
\description{inversebase function}
\details{A function to compute the base of the inverse os a block circulant matrix, given the base of the matrix}
\value{the base matrix of the inverse of the circulant matrix}
\arguments{\item{x}{the base matrix of a block circulant matrix}}
