\name{is.SPD}
\alias{is.SPD}
\title{is...}
\usage{is.SPD(base)}
\description{is.SPD function}
\details{A function to compute whether a block circulant matrix is symmetric positive definite (SPD), given its base matrix.}
\value{logical, whether the circulant matrix the base represents is SPD}
\arguments{\item{base}{base matrix of a block circulant matrix}}
