% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcIterator.R
\name{is.burnin}
\alias{is.burnin}
\title{is this a burn-in iteration?}
\usage{
is.burnin(obj)
}
\arguments{
\item{obj}{an mcmc iterator}
}
\value{
TRUE or FALSE
}
\description{
if this mcmc iteration is in the burn-in period, return TRUE
}
