% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{is.pow2}
\alias{is.pow2}
\title{is.pow2 function}
\usage{
is.pow2(num)
}
\arguments{
\item{num}{a numeric}
}
\value{
logical: is num a power of 2?
}
\description{
Tests whether a number id
}
\examples{
is.pow2(128)  # TRUE
is.pow2(64.9) # FALSE
}
