% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcIterator.R
\name{is.retain}
\alias{is.retain}
\title{do we retain this iteration?}
\usage{
is.retain(obj)
}
\arguments{
\item{obj}{an mcmc iterator}
}
\value{
TRUE or FALSE
}
\description{
if this mcmc iteration is one not thinned out, this is true
}
