% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialOnly.R
\name{is.SPD}
\alias{is.SPD}
\title{is.SPD function}
\usage{
is.SPD(base)
}
\arguments{
\item{base}{base matrix of a block circulant matrix}
}
\value{
logical, whether the circulant matrix the base represents is SPD
}
\description{
A function to compute whether a block circulant matrix is symmetric positive definite (SPD), given its base matrix.
}
