\name{iteration}
\alias{iteration}
\title{iteration number...}
\usage{iteration(obj)}
\description{iteration number}
\details{within a loop, this is the iteration number we are currently doing.

get the iteration number}
\value{integer iteration number, starting from 1.}
\arguments{\item{obj}{an mcmc iterator}}
