% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimumContrast.R
\name{K.u}
\alias{K.u}
\title{K.u function}
\usage{
K.u(u, phi, sig2, model, ...)
}
\arguments{
\item{u}{Spatial lag at which we wish to find the theoretical K function}

\item{phi}{Spatial scale parameter value}

\item{sig2}{Spatial variance parameter value}

\item{model}{A character string specifying the form of the theoretical spatial correlation function (matches 'model' argument for CovarianceFct in the RandomFields packages)}

\item{...}{Additional arguments if required for definition of the correlation function as per 'model'. See ?CovarianceFct (RandomFields)}
}
\value{
A single numeric value representing the theoretical K function evaluated at u.
}
\description{
A function to compute the theoretical K function for the LGCP.
}
