% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimumContrast.R
\name{K.val}
\alias{K.val}
\title{K.val function}
\usage{
K.val(val, phi, sig2, model, ...)
}
\arguments{
\item{val}{Spatial lag}

\item{phi}{Spatial scale parameter value}

\item{sig2}{Spatial variance parameter value}

\item{model}{A character string specifying the form of the theoretical spatial correlation function (matches 'model' argument for CovarianceFct in the RandomFields packages)}

\item{...}{Additional arguments if required for definition of the correlation function as per 'model'. See ?CovarianceFct (RandomFields)}
}
\value{
A single numeric value representing a component of the theoretical K function
}
\description{
An internal function used in computing the theoretical K function for the LGCP. See \link{K.u} for the theoretical K.
}
