\name{KinhomAverage}
\alias{KinhomAverage}
\title{KinhomAverage function...}
\usage{KinhomAverage(xyt, spatial.intensity, temporal.intensity,
    time.window=xyt$tlim, rvals, correction="iso",
    suppresswarnings=FALSE)}
\description{KinhomAverage function}
\details{A function to estimate the inhomogeneous K function for a spatiotemporal point process. The method of computation is similar to
\link{ginhomAverage}, see eq (8) Diggle P, Rowlingson B, Su T (2005) to see how this is computed.}
\value{time average of inhomogenous K function.}
\references{\enumerate{
\item Baddeley AJ, Moller J, Waagepetersen R (2000). Non-and semi-parametric estimation of interaction in inhomogeneous point patterns. Statistica Neerlandica, 54, 329-350.
\item Brix A, Diggle PJ (2001). Spatiotemporal Prediction for log-Gaussian Cox processes. Journal of the Royal Statistical Society, Series B, 63(4), 823-841.
\item Diggle P, Rowlingson B, Su T (2005). Point Process Methodology for On-line Spatio-temporal Disease Surveillance. Environmetrics, 16(5), 423-434.
}}
\seealso{\link{ginhomAverage}, \link{spatialparsEst}, \link{thetaEst}, \link{lambdaEst}, \link{muEst}}
\arguments{\item{xyt}{an object of class stppp}
\item{spatial.intensity}{A spatialAtRisk object}
\item{temporal.intensity}{A temporalAtRisk object}
\item{time.window}{time interval contained in the interval xyt$tlim over which to compute average. Useful if there is a lot of data over a lot of time points.}
\item{rvals}{Vector of values for the argument r at which the inhmogeneous K function should be evaluated (see ?Kinhom). There is a sensible default.}
\item{correction}{choice of edge correction to use, see ?Kinhom, default is Ripley isotropic correction}
\item{suppresswarnings}{Whether or not to suppress warnings generated by Kinhom}}
