\name{lambdaEst}
\alias{lambdaEst}
\title{lambdaEst function...}
\usage{lambdaEst(xyt, ...)}
\description{lambdaEst function}
\details{Generic function for estimating bivariate densities by eye. Specific methods exist for stppp objects and ppp objects.}
\value{method lambdaEst}
\seealso{\link{lambdaEst.stppp}, \link{lambdaEst.ppp}}
\arguments{\item{xyt}{an object}
\item{...}{additional arguments}}
