\name{lgcpgrid.array}
\alias{lgcpgrid.array}
\title{lgcpgrid...}
\usage{\method{lgcpgrid}{array} (grid, ...)}
\description{lgcpgrid.array function}
\details{Creates an lgcp grid object from an 3-dimensional array.}
\value{an object of class lgcpgrid}
\seealso{\link{lgcpgrid.list}, \link{as.list.lgcpgrid}, \link{print.lgcpgrid}, 
\link{summary.lgcpgrid}, \link{quantile.lgcpgrid}, \link{image.lgcpgrid}, \link{plot.lgcpgrid}}
\arguments{\item{grid}{a three dimensional array object}
\item{...}{other arguments}}
