% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{lgcpgrid.array}
\alias{lgcpgrid.array}
\title{lgcpgrid.array function}
\usage{
\method{lgcpgrid}{array}(
  grid,
  xvals = 1:dim(grid)[1],
  yvals = 1:dim(grid)[2],
  zvals = 1:dim(grid)[3],
  ...
)
}
\arguments{
\item{grid}{a three dimensional array object}

\item{xvals}{optional vector of x-coordinates associated to grid. By default, this is the cell index in the x direction.}

\item{yvals}{optional vector of y-coordinates associated to grid. By default, this is the cell index in the y direction.}

\item{zvals}{optional vector of z-coordinates (time) associated to grid. By default, this is the cell index in the z direction.}

\item{...}{other arguments}
}
\value{
an object of class lgcpgrid
}
\description{
Creates an lgcp grid object from an 3-dimensional array.
}
\seealso{
\link{lgcpgrid.list}, \link{as.list.lgcpgrid}, \link{print.lgcpgrid}, 
\link{summary.lgcpgrid}, \link{quantile.lgcpgrid}, \link{image.lgcpgrid}, \link{plot.lgcpgrid}
}
