\name{lgcpgrid.matrix}
\alias{lgcpgrid.matrix}
\title{lgcpgrid...}
\usage{\method{lgcpgrid}{matrix} (grid, ...)}
\description{lgcpgrid.matrix function}
\details{Creates an lgcp grid object from an 2-dimensional matrix.}
\value{an object of class lgcpgrid}
\seealso{\link{lgcpgrid.list}, \link{as.list.lgcpgrid}, \link{print.lgcpgrid}, 
\link{summary.lgcpgrid}, \link{quantile.lgcpgrid}, \link{image.lgcpgrid}, \link{plot.lgcpgrid}}
\arguments{\item{grid}{a three dimensional array object}
\item{...}{other arguments}}
