\name{lgcpgrid}
\alias{lgcpgrid}
\title{lgcpgrid function...}
\usage{lgcpgrid(grid, ...)}
\description{lgcpgrid function}
\details{Generic function for the hadling of list objects where each element of the list is
a matrix. Each matrix is assumed to have the same dimension. Such objects arise from the 
various routines in the package lgcp.}
\value{method lgcpgrid}
\arguments{\item{grid}{a list object with each member of the list being a numeric matrix, each matrix having the same dimension}
\item{...}{other arguments}}
