% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultitypeSpatialOnlyPlusParameters.R
\name{lgcpSimMultitypeSpatialCovariates}
\alias{lgcpSimMultitypeSpatialCovariates}
\title{lgcpSimMultitypeSpatialCovariates function}
\usage{
lgcpSimMultitypeSpatialCovariates(
  formulaList,
  owin,
  regionalcovariates,
  pixelcovariates,
  betaList,
  spatial.offsetList = NULL,
  cellwidth,
  model.parameters,
  spatial.covmodel = "exponential",
  covpars = c(),
  ext = 2,
  plot = FALSE,
  inclusion = "touching"
)
}
\arguments{
\item{formulaList}{a list of formulae objetcs}

\item{owin}{a spatstat owin object on which to simulate the data}

\item{regionalcovariates}{a SpatialPolygonsDataFrame object}

\item{pixelcovariates}{a SpatialPixelsDataFrame object}

\item{betaList}{list of beta parameters}

\item{spatial.offsetList}{list of poisson offsets}

\item{cellwidth}{cellwidth}

\item{model.parameters}{model parameters, a list eg list(sigma=1,phi=0.2)}

\item{spatial.covmodel}{the choice of spatial covariance model, can be anything from the RandomFields covariance function, CovariacenFct.}

\item{covpars}{additional covariance parameters, for the chosen model, optional.}

\item{ext}{number of times to extend the simulation window}

\item{plot}{whether to plot the results automatically}

\item{inclusion}{criterion for cells being included into observation window. Either 'touching' or 'centroid'. The former, the default, includes all cells that touch the observation window, the latter includes all cells whose centroids are inside the observation window.}
}
\value{
a marked ppp object, the simulated data
}
\description{
A function to Simulate multivariate point process models
}
