% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{LogGaussianPrior}
\alias{LogGaussianPrior}
\title{LogGaussianPrior function}
\usage{
LogGaussianPrior(mean, variance)
}
\arguments{
\item{mean}{a vector of length 2 representing the mean (on the log scale)}

\item{variance}{a 2x2 matrix representing the variance (on the log scale)}
}
\value{
an object of class LogGaussianPrior that can be passed to the function PriorSpec.
}
\description{
A function to create a Gaussian prior on the log scale
}
\examples{
\dontrun{LogGaussianPrior(mean=log(c(1,500)),variance=diag(0.15,2))}
}
\seealso{
\link{GaussianPrior}, link{PriorSpec.list}
}
