% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{ltar}
\alias{ltar}
\title{ltar function}
\usage{
ltar(lg)
}
\arguments{
\item{lg}{an object produced by a call to lgcpPredictSpatialPlusPars, lgcpPredictAggregateSpatialPlusPars, lgcpPredictSpatioTemporalPlusPars orlgcpPredictMultitypeSpatialPlusPars}
}
\value{
the log-target from each saved iteration of the MCMC chain.
}
\description{
A function to return the sampled log-target from a call to the function lgcpPredictSpatialPlusPars, lgcpPredictAggregateSpatialPlusPars,
lgcpPredictSpatioTemporalPlusPars or lgcpPredictMultitypeSpatialPlusPars. This is used as a convergence diagnostic.
}
\seealso{
\link{autocorr}, \link{parautocorr}, \link{traceplots}, \link{parsummary}, \link{textsummary},
\link{priorpost}, \link{postcov}, \link{exceedProbs}, \link{betavals}, \link{etavals}
}
