% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultitypeSpatialOnlyPlusParameters.R
\name{MALAlgcpMultitypeSpatial.PlusPars}
\alias{MALAlgcpMultitypeSpatial.PlusPars}
\title{MALAlgcpMultitypeSpatial.PlusPars function}
\usage{
MALAlgcpMultitypeSpatial.PlusPars(
  mcmcloop,
  inits,
  adaptivescheme,
  M,
  N,
  Mext,
  Next,
  mcens,
  ncens,
  formulaList,
  zml,
  Zmat,
  model.priorsList,
  model.initsList,
  fftgrid,
  spatial.covmodelList,
  nis,
  cellarea,
  spatialvals,
  cellInside,
  MCMCdiag,
  gradtrunc,
  gridfun,
  gridav,
  marks,
  ntypes,
  d
)
}
\arguments{
\item{mcmcloop}{details of the mcmc loop}

\item{inits}{initial values}

\item{adaptivescheme}{the adaptive MCMC scheme}

\item{M}{number of grid cells in x direction}

\item{N}{number of grid cells in y direction}

\item{Mext}{number of extended grid cells in x direction}

\item{Next}{number of extended grid cells in y direction}

\item{mcens}{centroids in x direction}

\item{ncens}{centroids in y direction}

\item{formulaList}{a list of formula objects of the form X ~ var1 + var2 etc.}

\item{zml}{list of design matrices}

\item{Zmat}{a design matrix constructed using getZmat}

\item{model.priorsList}{list of model priors, see lgcpPriors}

\item{model.initsList}{list of model initial values, see lgcpInits}

\item{fftgrid}{an objects of class FFTgrid, see genFFTgrid}

\item{spatial.covmodelList}{list of spatial covariance models constructed using CovFunction}

\item{nis}{cell counts on the etended grid}

\item{cellarea}{the cell area}

\item{spatialvals}{inerpolated poisson offset on fft grid}

\item{cellInside}{0-1 matrix indicating inclusion in the observation window}

\item{MCMCdiag}{not used}

\item{gradtrunc}{gradient truncation parameter}

\item{gridfun}{used to specify other actions to be taken, e.g. dumping MCMC output to disk.}

\item{gridav}{used for computing Monte Carlo expectations online}

\item{marks}{the marks from the marked ppp object}

\item{ntypes}{the number of types being analysed}

\item{d}{matrix of toral distances}
}
\value{
output from the MCMC run
}
\description{
A function to run the MCMC algorithm for multivariate spatial point process data. Not for general purpose use.
}
