% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpPredictSpatial.R
\name{MALAlgcpSpatial}
\alias{MALAlgcpSpatial}
\title{MALAlgcpSpatial function}
\usage{
MALAlgcpSpatial(
  mcmcloop,
  inits,
  adaptivescheme,
  M,
  N,
  Mext,
  Next,
  sigma,
  phi,
  mu,
  nis,
  cellarea,
  spatialvals,
  scaleconst,
  rootQeigs,
  invrootQeigs,
  cellInside,
  MCMCdiag,
  gradtrunc,
  gridfun,
  gridav,
  mcens,
  ncens
)
}
\arguments{
\item{mcmcloop}{an mcmcLoop object}

\item{inits}{initial values from mcmc.control}

\item{adaptivescheme}{adaptive scheme from mcmc.control}

\item{M}{number of cells in x direction on output grid}

\item{N}{number of cells in y direction on output grid}

\item{Mext}{number of cells in x direction on extended output grid}

\item{Next}{number of cells in y direction on extended output grid}

\item{sigma}{spatial covariance parameter sigma}

\item{phi}{spatial covariance parameter phi}

\item{mu}{spatial covariance parameter mu}

\item{nis}{cell counts matrix}

\item{cellarea}{area of cells}

\item{spatialvals}{spatial at risk, function lambda, interpolated onto the requisite grid}

\item{scaleconst}{expected number of observations}

\item{rootQeigs}{square root of eigenvalues of precision matrix}

\item{invrootQeigs}{inverse square root of eigenvalues of precision matrix}

\item{cellInside}{logical matrix dictating whether cells are inside the observation window}

\item{MCMCdiag}{defunct}

\item{gradtrunc}{gradient truncation parameter}

\item{gridfun}{grid functions}

\item{gridav}{grid average functions}

\item{mcens}{x-coordinates of cell centroids}

\item{ncens}{y-coordinates of cell centroids}
}
\value{
object passed back to lgcpPredictSpatial
}
\description{
ADVANCED USE ONLY A function to perform MALA for the spatial only case
}
