% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcp-INLA.R
\name{matchcovariance}
\alias{matchcovariance}
\title{matchcovariance function}
\usage{
matchcovariance(
  xg,
  yg,
  ns,
  sigma,
  phi,
  model,
  additionalparameters,
  verbose = TRUE,
  r = 1,
  method = "Nelder-Mead"
)
}
\arguments{
\item{xg}{x grid must be equally spaced}

\item{yg}{y grid must be equally spaced}

\item{ns}{neighbourhood size}

\item{sigma}{spatial variability parameter}

\item{phi}{spatial dependence parameter}

\item{model}{covariance model, see ?CovarianceFct}

\item{additionalparameters}{additional parameters for chosen covariance model}

\item{verbose}{whether or not to print stuff generated by the optimiser}

\item{r}{parameter used in optimisation, see Rue and Held (2005) pp 188. default value 1.}

\item{method}{The choice of optimising routine must either be 'Nelder-Mead' or 'BFGS'. see ?optim}
}
\value{
...
}
\description{
A function to match the covariance matrix of a Gaussian Field with an approximate GMRF with neighbourhood size ns.
}
