% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{maternCovFct15}
\alias{maternCovFct15}
\title{maternCovFct15 function}
\usage{
maternCovFct15(d, CovParameters)
}
\arguments{
\item{d}{toral distance}

\item{CovParameters}{parameters of the latent field, an object of class "CovParamaters".}
}
\value{
the exponential covariance function
}
\description{
A function to declare and also evaluate an Matern 1.5 covariance function.
}
\seealso{
\link{CovFunction.function}, \link{RandomFieldsCovFct}, \link{SpikedExponentialCovFct}
}
\author{
Dominic Schumacher
}
