\name{mcmcProgressTextBar}
\alias{mcmcProgressTextBar}
\title{text bar progress monitor...}
\usage{mcmcProgressTextBar(mcmcloop)}
\description{text bar progress monitor}
\details{a progress monitor that uses a text progress bar}
\value{a progress monitor}
\arguments{\item{mcmcloop}{an mcmc loop iterator}}
