% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcProgress.R
\name{mcmcProgressTk}
\alias{mcmcProgressTk}
\title{graphical progress monitor}
\usage{
mcmcProgressTk(mcmcloop)
}
\arguments{
\item{mcmcloop}{an mcmc loop iterator}
}
\value{
a progress monitor
}
\description{
a progress monitor that uses tcltk dialogs
}
