\name{mcmctrace.lgcpPredict}
\alias{mcmctrace.lgcpPredict}
\title{mcmctrace...}
\usage{\method{mcmctrace}{lgcpPredict} (obj, ...)}
\description{mcmctrace.lgcpPredict function}
\details{If \code{MCMCdiag} was positive when \code{lgcpPredict} was called, then this retrieves information from the chains stored.}
\value{returns the saved MCMC chains in an object of class \code{mcmcdiag}.}
\seealso{\link{lgcpPredict}, \link{plot.mcmcdiag}}
\arguments{\item{obj}{an object of class lgcpPredict}
\item{...}{additional arguments}}
