\name{mcmctrace}
\alias{mcmctrace}
\title{mcmctrace function...}
\usage{mcmctrace(obj, ...)}
\description{mcmctrace function}
\details{Generic function to extract the information required to produce MCMC trace plots.}
\value{method mcmctrace}
\arguments{\item{obj}{an object}
\item{...}{additional arguments}}
