\name{meanfield.lgcpPredict}
\alias{meanfield.lgcpPredict}
\title{meanfield...}
\usage{\method{meanfield}{lgcpPredict} (obj, ...)}
\description{meanfield.lgcpPredict function}
\details{This is an accessor function for objects of class \code{lgcpPredict} and returns the mean of the 
field Y as an lgcpgrid object.}
\value{returns the cell-wise mean of Y computed via Monte Carlo.}
\seealso{\link{lgcpPredict}, \link{lgcpgrid}}
\arguments{\item{obj}{an object of class lgcpPredict}
\item{...}{additional arguments}}
