\name{meanfield.lgcpPredictINLA}
\alias{meanfield.lgcpPredictINLA}
\title{meanfield...}
\usage{\method{meanfield}{lgcpPredictINLA} (obj, ...)}
\description{meanfield.lgcpPredictINLA function}
\details{A function to return the mean of the latent field from a call to lgcpPredictINLA output.}
\value{the mean of the latent field}
\arguments{\item{obj}{an object of class lgcpPredictINLA}
\item{...}{other arguments}}
