% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{meanfield}
\alias{meanfield}
\title{meanfield function}
\usage{
meanfield(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{additional arguments}
}
\value{
method meanfield
}
\description{
Generic function to extract the mean of the latent field Y.
}
