\name{mstppp.list}
\alias{mstppp.list}
\title{mstppp...}
\usage{\method{mstppp}{list} (P, ...)}
\description{mstppp.list function}
\details{Construct a marked space-time planar point pattern from a list object}
\value{an object of class mstppp}
\seealso{\link{mstppp}, \link{mstppp.ppp},}
\arguments{\item{P}{list object containing $xyt, an (n x 3) matrix corresponding to (x,y,t) values; $tlim, a vector of length 2 givign the observation time window, 
$window giving an owin spatial observation winow, see ?owin for more details, and $data, a data frame containing the collection of marks}
\item{...}{additional arguments}}
