\name{mstppp.ppp}
\alias{mstppp.ppp}
\title{mstppp...}
\usage{\method{mstppp}{ppp} (P, t, tlim, data, ...)}
\description{mstppp.ppp function}
\details{Construct a marked space-time planar point pattern from a ppp object}
\value{an object of class mstppp}
\seealso{\link{mstppp}, \link{mstppp.list}}
\arguments{\item{P}{a spatstat ppp object}
\item{t}{a vector of length P$n}
\item{tlim}{a vector of length 2 specifying the observation time window}
\item{data}{a data frame containing the collection of marks}
\item{...}{additional arguments}}
