\name{mstppp.stppp}
\alias{mstppp.stppp}
\title{mstppp...}
\usage{\method{mstppp}{stppp} (P, data, ...)}
\description{mstppp.stppp function}
\details{Construct a marked space-time planar point pattern from an stppp object}
\value{an object of class mstppp}
\seealso{\link{mstppp}, \link{mstppp.list}}
\arguments{\item{P}{an lgcp stppp object}
\item{data}{a data frame containing the collection of marks}
\item{...}{additional arguments}}
