% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstpppClassDef.R
\name{mstppp.list}
\alias{mstppp.list}
\title{mstppp.list function}
\usage{
\method{mstppp}{list}(P, ...)
}
\arguments{
\item{P}{list object containing $xyt, an (n x 3) matrix corresponding to (x,y,t) values; $tlim, a vector of length 2 givign the observation time window, 
$window giving an owin spatial observation winow, see ?owin for more details, and $data, a data frame containing the collection of marks}

\item{...}{additional arguments}
}
\value{
an object of class mstppp
}
\description{
Construct a marked space-time planar point pattern from a list object
}
\seealso{
\link{mstppp}, \link{mstppp.ppp},
}
