% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstpppClassDef.R
\name{mstppp.ppp}
\alias{mstppp.ppp}
\title{mstppp.ppp function}
\usage{
\method{mstppp}{ppp}(P, t, tlim, data, ...)
}
\arguments{
\item{P}{a spatstat ppp object}

\item{t}{a vector of length P$n}

\item{tlim}{a vector of length 2 specifying the observation time window}

\item{data}{a data frame containing the collection of marks}

\item{...}{additional arguments}
}
\value{
an object of class mstppp
}
\description{
Construct a marked space-time planar point pattern from a ppp object
}
\seealso{
\link{mstppp}, \link{mstppp.list}
}
