\name{mstppp}
\alias{mstppp}
\title{mstppp function...}
\usage{mstppp(P, ...)}
\description{mstppp function}
\details{Generic function used in the construction of marked space-time planar point patterns. An mstppp object is like an stppp object,
but with an extra component containing a data frame (the mark information).

Observations are assumed to occur in the plane and the observation window is assumed not to change over time.}
\value{method mstppp}
\seealso{\link{mstppp}, \link{mstppp.ppp}, \link{mstppp.list}}
\arguments{\item{P}{an object}
\item{...}{additional arguments}}
