\name{multiply.list}
\alias{multiply.list}
\title{multiply...}
\usage{multiply.list(list1, list2)}
\description{multiply.list function}
\details{This function multiplies the elements of two list objects together and returns the result in another list object.}
\value{a list with ith entry the sum of list1[[i]] and list2[[i]]}
\arguments{\item{list1}{a list of objects that could be summed using "+"}
\item{list2}{a list of objects that could be summed using "+"}}
