\name{neigh2D}
\alias{neigh2D}
\title{neigh2D function...}
\usage{neigh2D(i, j, ns, M, N)}
\description{neigh2D function}
\details{A function to compute the neighbours of a cell on a toral grid}
\value{the cell indices of the neighbours}
\arguments{\item{i}{cell index i}
\item{j}{cell index j}
\item{ns}{number of neighbours either side}
\item{M}{size of grid in x direction}
\item{N}{size of grid in y direction}}
