\name{nullFunction}
\alias{nullFunction}
\title{nullFunction function...}
\description{nullFunction function}
\details{This is a null function and performs no action.}
\value{object of class nullFunction}
\seealso{\link{setoutput}, \link{GFinitialise}, \link{GFupdate}, \link{GFfinalise}, \link{GFreturnvalue}}
