\name{osppp2merc}
\alias{osppp2merc}
\title{osppp2merc function}
\usage{
  osppp2merc(obj)
}
\arguments{
  \item{obj}{a ppp object in OSGB}
}
\value{
  a ppp object in Mercator
}
\description{
  A function to transform a ppp object in the OS GB
  projection (epsg:27700) to a ppp object in the Mercator
  (epsg:3857) projection.
}

