% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{parautocorr}
\alias{parautocorr}
\title{parautocorr function}
\usage{
parautocorr(obj, xlab = "Lag", ylab = NULL, main = "", ask = TRUE, ...)
}
\arguments{
\item{obj}{an object produced by a call to lgcpPredictSpatialPlusPars, lgcpPredictAggregateSpatialPlusPars, lgcpPredictSpatioTemporalPlusPars orlgcpPredictMultitypeSpatialPlusPars}

\item{xlab}{optional label for x-axis, there is a sensible default.}

\item{ylab}{optional label for y-axis, there is a sensible default.}

\item{main}{optional title of the plot, there is a sensible default.}

\item{ask}{the paramter "ask", see ?par}

\item{...}{other arguments passed to the function "hist"}
}
\value{
produces autocorrelation plots of the parameters beta and eta
}
\description{
A function to produce autocorrelation plots for the paramerers beta and eta from a call to the function lgcpPredictSpatialPlusPars, lgcpPredictAggregateSpatialPlusPars, lgcpPredictSpatioTemporalPlusPars or lgcpPredictMultitypeSpatialPlusPars
}
\seealso{
\link{ltar}, \link{autocorr}, \link{traceplots}, \link{parsummary}, \link{textsummary},
\link{priorpost}, \link{postcov}, \link{exceedProbs}, \link{betavals}, \link{etavals}
}
