% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{parsummary}
\alias{parsummary}
\title{parsummary function}
\usage{
parsummary(obj, expon = TRUE, LaTeX = FALSE, ...)
}
\arguments{
\item{obj}{an object produced by a call to lgcpPredictSpatialPlusPars, lgcpPredictAggregateSpatialPlusPars, lgcpPredictSpatioTemporalPlusPars orlgcpPredictMultitypeSpatialPlusPars}

\item{expon}{whether to exponentiate the results, so that the parameters beta haev the interpretation of "relative risk per unit increase in the covariate" default is TRUE}

\item{LaTeX}{whether to print paramter names using LaTeX symbols (if the table is later to be exported to a LaTeX document)}

\item{...}{other arguments}
}
\value{
a data frame containing the median, 0.025 and 0.975 quantiles.
}
\description{
A function to produce a summary table for the parameters beta and eta from a call to the function lgcpPredictSpatialPlusPars, lgcpPredictAggregateSpatialPlusPars, lgcpPredictSpatioTemporalPlusPars or lgcpPredictMultitypeSpatialPlusPars
}
\seealso{
\link{ltar}, \link{autocorr}, \link{parautocorr}, \link{traceplots}, \link{textsummary},
\link{priorpost}, \link{postcov}, \link{exceedProbs}, \link{betavals}, \link{etavals}
}
