\name{plot.fromSPDF}
\alias{plot.fromSPDF}
\title{plot...}
\usage{\method{plot}{fromSPDF} (x, ...)}
\description{plot.fromSPDF function}
\details{Plot method for objects of class fromSPDF.}
\value{prints the object}
\arguments{\item{x}{an object of class spatialAtRisk}
\item{...}{additional arguments}}
