\name{plot.mstppp}
\alias{plot.mstppp}
\title{plot...}
\usage{\method{plot}{mstppp} (x, cols="red", ...)}
\description{plot.mstppp function}
\details{Plot method for mstppp objects}
\value{plots the mstppp object x}
\arguments{\item{x}{an object of class mstppp}
\item{cols}{optional vector of colours to plot points with}
\item{...}{additional arguments passed to plot}}
