\name{plot.temporalAtRisk}
\alias{plot.temporalAtRisk}
\title{plot...}
\usage{\method{plot}{temporalAtRisk} (x, ...)}
\description{plot.temporalAtRisk function}
\details{Pot a temporalAtRisk object.}
\value{print the object}
\seealso{\link{temporalAtRisk}, \link{spatialAtRisk}, \link{temporalAtRisk.numeric}, \link{temporalAtRisk.function}, \link{constantInTime}, \link{constantInTime.numeric}, \link{constantInTime.stppp}, \link{print.temporalAtRisk},}
\arguments{\item{x}{an object}
\item{...}{additional arguments}}
