\name{plot.fromSPDF}
\alias{plot.fromSPDF}
\title{plot.fromSPDF function}
\usage{
  \method{plot}{fromSPDF} (x, ...)
}
\arguments{
  \item{x}{an object of class spatialAtRisk}

  \item{...}{additional arguments}
}
\value{
  prints the object
}
\description{
  Plot method for objects of class fromSPDF.
}

