% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stpppClassDef.R
\name{plot.stppp}
\alias{plot.stppp}
\title{plot.stppp function}
\usage{
\method{plot}{stppp}(x, ...)
}
\arguments{
\item{x}{an object of class stppp}

\item{...}{additional arguments passed to plot}
}
\value{
plots the stppp object x
}
\description{
Plot method for stppp objects
}
