% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporalAtRiskClassDef.R
\name{plot.temporalAtRisk}
\alias{plot.temporalAtRisk}
\title{plot.temporalAtRisk function}
\usage{
\method{plot}{temporalAtRisk}(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{additional arguments}
}
\value{
print the object
}
\description{
Pot a temporalAtRisk object.
}
\seealso{
\link{temporalAtRisk}, \link{spatialAtRisk}, \link{temporalAtRisk.numeric}, \link{temporalAtRisk.function}, \link{constantInTime}, \link{constantInTime.numeric}, \link{constantInTime.stppp}, \link{print.temporalAtRisk},
}
