\name{plotExceed}
\alias{plotExceed}
\title{plotExceed function...}
\usage{plotExceed(obj, ...)}
\description{plotExceed function}
\details{A generic function for plotting exceedance probabilities.}
\value{generic function returning method plotExceed}
\seealso{\link{plotExceed.lgcpPredict}, \link{plotExceed.array}}
\arguments{\item{obj}{an object}
\item{...}{additional arguments}}
