% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{plotit}
\alias{plotit}
\title{plotit function}
\usage{
plotit(x)
}
\arguments{
\item{x}{an a list, matrix, or GPrealisation object.}
}
\value{
plots the objects.
}
\description{
A function to plot various objects. A developmental tool: not intended for general use
}
