% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{postcov}
\alias{postcov}
\title{postcov function}
\usage{
postcov(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{additional arguments}
}
\value{
method postcov
}
\description{
Generic function for producing plots of the posterior covariance function from a call to the function lgcpPredictSpatialPlusPars, lgcpPredictAggregateSpatialPlusPars,
lgcpPredictSpatioTemporalPlusPars or lgcpPredictMultitypeSpatialPlusPars.
}
\seealso{
\link{postcov.lgcpPredictSpatialOnlyPlusParameters},\link{postcov.lgcpPredictAggregateSpatialPlusParameters}, \link{postcov.lgcpPredictSpatioTemporalPlusParameters}, \link{postcov.lgcpPredictMultitypeSpatialPlusParameters},
\link{ltar}, \link{autocorr}, \link{parautocorr}, \link{traceplots}, \link{parsummary}, \link{textsummary},
\link{priorpost}, \link{exceedProbs}, \link{betavals}, \link{etavals}
}
