\name{print.fromFunction}
\alias{print.fromFunction}
\title{print...}
\usage{\method{print}{fromFunction} (x, ...)}
\description{print.fromFunction function}
\details{Print method for objects of class fromFunction.}
\value{prints the object}
\arguments{\item{x}{an object of class spatialAtRisk}
\item{...}{additional arguments}}
