\name{print.fromSPDF}
\alias{print.fromSPDF}
\title{print...}
\usage{\method{print}{fromSPDF} (x, ...)}
\description{print.fromSPDF function}
\details{Print method for objects of class fromSPDF.}
\value{prints the object}
\arguments{\item{x}{an object of class spatialAtRisk}
\item{...}{additional arguments}}
