\name{print.fromXYZ}
\alias{print.fromXYZ}
\title{print...}
\usage{\method{print}{fromXYZ} (x, ...)}
\description{print.fromXYZ function}
\details{Print method for objects of class fromXYZ.}
\value{prints the object}
\arguments{\item{x}{an object of class spatialAtRisk}
\item{...}{additional arguments}}
