\name{print.dump2dir}
\alias{print.dump2dir}
\title{print...}
\usage{\method{print}{dump2dir} (x, ...)}
\description{print.dump2dir function}
\details{Display function for dump2dir objects.}
\value{nothing}
\seealso{\link{dump2dir},}
\arguments{\item{x}{an object of class dump2dir}
\item{...}{additional arguments}}
