\name{print.fromFunction}
\alias{print.fromFunction}
\title{print.fromFunction function}
\usage{
\method{print}{fromFunction}(x, ...)
}
\arguments{
  \item{x}{an object of class spatialAtRisk}

  \item{...}{additional arguments}
}
\value{
prints the object
}
\description{
Print method for objects of class fromFunction.
}

