\name{print.gridaverage}
\alias{print.gridaverage}
\title{print...}
\usage{\method{print}{gridaverage} (x, ...)}
\description{print.gridaverage function}
\details{Print method for \code{gridaverage} objects}
\value{just prints out details}
\arguments{\item{x}{an object of class gridaverage}
\item{...}{other arguments}}
