\name{print.lgcpgrid}
\alias{print.lgcpgrid}
\title{print...}
\usage{\method{print}{lgcpgrid} (x, ...)}
\description{print.lgcpgrid function}
\details{Print method for lgcp grid objects.}
\value{just prints out details to the console}
\seealso{\link{lgcpgrid.list}, \link{lgcpgrid.array}, \link{as.list.lgcpgrid},
\link{summary.lgcpgrid} \link{quantile.lgcpgrid} \link{image.lgcpgrid} \link{plot.lgcpgrid}}
\arguments{\item{x}{an object of class lgcpgrid}
\item{...}{other arguments}}
