\name{print.lgcpPredict}
\alias{print.lgcpPredict}
\title{print...}
\usage{\method{print}{lgcpPredict} (x, ...)}
\description{print.lgcpPredict function}
\details{Print method for lgcpPredict objects.}
\value{just prints information to the screen}
\seealso{\link{lgcpPredict}}
\arguments{\item{x}{an object of class dump2dir}
\item{...}{additional arguments}}
