% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcIterator.R
\name{print.mcmc}
\alias{print.mcmc}
\title{print.mcmc function}
\usage{
\method{print}{mcmc}(x, ...)
}
\arguments{
\item{x}{a mcmc iterator}

\item{...}{other args}
}
\description{
print method
print an mcmc iterator's details
}
