% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstpppClassDef.R
\name{print.mstppp}
\alias{print.mstppp}
\title{print.mstppp function}
\usage{
\method{print}{mstppp}(x, ...)
}
\arguments{
\item{x}{an object of class mstppp}

\item{...}{additional arguments}
}
\value{
prints the mstppp object x
}
\description{
Print method for mstppp objects
}
