\name{print.stapp}
\alias{print.stapp}
\title{print...}
\usage{\method{print}{stapp} (x, printhead=TRUE, ...)}
\description{print.stapp function}
\details{Print method for stapp objects}
\value{prints the stapp object x}
\arguments{\item{x}{an object of class stapp}
\item{printhead}{whether or not to print the head of the counts matrix}
\item{...}{additional arguments}}
