% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stpppClassDef.R
\name{print.stppp}
\alias{print.stppp}
\title{print.stppp function}
\usage{
\method{print}{stppp}(x, ...)
}
\arguments{
\item{x}{an object of class stppp}

\item{...}{additional arguments}
}
\value{
prints the stppp object x
}
\description{
Print method for stppp objects
}
