\name{print.temporalAtRisk}
\alias{print.temporalAtRisk}
\title{print.temporalAtRisk function}
\usage{
\method{print}{temporalAtRisk}(x, ...)
}
\arguments{
  \item{x}{an object}

  \item{...}{additional arguments}
}
\value{
print the object
}
\description{
Printing method for temporalAtRisk objects.
}
\seealso{
\link{temporalAtRisk}, \link{spatialAtRisk},
\link{temporalAtRisk.numeric},
\link{temporalAtRisk.function}, \link{constantInTime},
\link{constantInTime.numeric}, \link{constantInTime.stppp},
\link{plot.temporalAtRisk}
}

