\name{quantile.lgcpgrid}
\alias{quantile.lgcpgrid}
\title{quantile...}
\usage{\method{quantile}{lgcpgrid} (x, ...)}
\description{quantile.lgcpgrid function}
\details{Quantile method for lgcp objects. This just applies the quantile function to each of
the elements of x$grid}
\value{Quantiles per grid, see ?quantile for further options}
\seealso{\link{lgcpgrid.list}, \link{lgcpgrid.array}, \link{as.list.lgcpgrid}, \link{print.lgcpgrid}, 
\link{summary.lgcpgrid}, \link{image.lgcpgrid}, \link{plot.lgcpgrid}}
\arguments{\item{x}{an object of class lgcpgrid}
\item{...}{other arguments}}
