% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{RandomFieldsCovFct}
\alias{RandomFieldsCovFct}
\title{RandomFieldsCovFct function}
\usage{
RandomFieldsCovFct(model, additionalparameters = c())
}
\arguments{
\item{model}{the choice of model e.g.  "matern"}

\item{additionalparameters}{additional parameters for chosen covariance model. See ?CovarianceFct}
}
\value{
a covariance function from the RandomFields package
}
\description{
A function to declare and also evaluate an covariance function from the RandomFields Package. See ?CovarianceFct. Note that the present version of
lgcp only offers estimation for sigma and phi, any additional paramters are treated as fixed.
}
\examples{
\dontrun{RandomFieldsCovFct(model="matern",additionalparameters=1)}
}
\seealso{
\link{CovFunction.function}, \link{exponentialCovFct}, \link{SpikedExponentialCovFct}, \link{CovarianceFct}
}
