\name{raster.lgcpgrid}
\alias{raster.lgcpgrid}
\title{raster...}
\usage{\method{raster}{lgcpgrid} (x, crs=NA, transpose=FALSE, ...)}
\description{raster.lgcpgrid function}
\details{A function to convert lgcpgrid objects into either a raster object, or a RasterBrick object.}
\value{...}
\arguments{\item{x}{an lgcpgrid object}
\item{crs}{PROJ4 type description of a map projection (optional). See ?raster}
\item{transpose}{Logical. Transpose the data? See ?brick method for array}
\item{...}{additional arguments}}
