% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{raster.lgcpgrid}
\alias{raster.lgcpgrid}
\title{raster.lgcpgrid function}
\usage{
\method{raster}{lgcpgrid}(x, crs = NA, transpose = FALSE, ...)
}
\arguments{
\item{x}{an lgcpgrid object}

\item{crs}{PROJ4 type description of a map projection (optional). See ?raster}

\item{transpose}{Logical. Transpose the data? See ?brick method for array}

\item{...}{additional arguments}
}
\value{
...
}
\description{
A function to convert lgcpgrid objects into either a raster object, or a RasterBrick object.
}
