\name{rescale.mstppp}
\alias{rescale.mstppp}
\title{rescale...}
\usage{\method{rescale}{mstppp} (X, s)}
\description{rescale.mstppp function}
\details{Rescale an mstppp object. Similar to rescale.ppp}
\value{a ppp object without observation times}
\arguments{\item{X}{an object of class mstppp}
\item{s}{scale as in rescale.ppp: x and y coordinaes are scaled by 1/s}}
