\name{rescale.mstppp}
\alias{rescale.mstppp}
\title{rescale.mstppp function}
\usage{
  \method{rescale}{mstppp} (X, s)
}
\arguments{
  \item{X}{an object of class mstppp}

  \item{s}{scale as in rescale.ppp: x and y coordinaes are
  scaled by 1/s}
}
\value{
  a ppp object without observation times
}
\description{
  Rescale an mstppp object. Similar to rescale.ppp
}

