\name{rescale.stppp}
\alias{rescale.stppp}
\title{rescale...}
\usage{\method{rescale}{stppp} (X, s)}
\description{rescale.stppp function}
\details{Rescale an stppp object. Similar to rescale.ppp}
\value{a ppp object without observation times}
\arguments{\item{X}{an object of class stppp}
\item{s}{scale as in rescale.ppp: x and y coordinaes are scaled by 1/s}}
