\name{roteffgain}
\alias{roteffgain}
\title{roteffgain function...}
\usage{roteffgain(xyt, cellwidth)}
\description{roteffgain function}
\details{Compute whether there might be any advantage in rotating the observation window in the object xyt for a proposed cell width.}
\value{whether or not there woud be any efficiency gain in the MALA by rotating window}
\seealso{\link{getRotation.stppp}}
\arguments{\item{xyt}{an object of class stppp}
\item{cellwidth}{size of grid on which to do MALA}}
