\name{rotmat}
\alias{rotmat}
\title{rotmat function...}
\usage{rotmat(theta)}
\description{rotmat function}
\details{This function returns a rotation matrix corresponding to an anticlockwise rotation of theta radians about the origin}
\value{the transformation matrix corresponding to an anticlockwise rotation of theta radians about the origin}
\arguments{\item{theta}{an angle in radians}}
