\name{rr.lgcpPredict}
\alias{rr.lgcpPredict}
\title{rr...}
\usage{\method{rr}{lgcpPredict} (obj, ...)}
\description{rr.lgcpPredict function}
\details{Accessor function returning the relative risk = exp(Y) as an lgcpgrid object.}
\value{the relative risk as computed my MCMC}
\seealso{\link{lgcpPredict}}
\arguments{\item{obj}{an lgcpPredict object}
\item{...}{additional arguments}}
