% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{samplePosterior}
\alias{samplePosterior}
\title{samplePosterior function}
\usage{
samplePosterior(x)
}
\arguments{
\item{x}{an object of class lgcpPredictSpatialOnlyPlusParameters or lgcpPredictAggregateSpatialPlusParameters}
}
\value{
a sample from the posterior named list object with names elements "eta", "beta" and "Y".
}
\description{
A function to draw a sample from the posterior of a spatial LGCP. Randomly selects an index i, and returns the ith value of eta,
the ith value of beta and the ith value of Y as a named list.
}
