\name{seintens.lgcpPredict}
\alias{seintens.lgcpPredict}
\title{seintens...}
\usage{\method{seintens}{lgcpPredict} (obj, ...)}
\description{seintens.lgcpPredict function}
\details{Accessor function returning the  standard error of the Poisson intensity as an lgcpgrid object.}
\value{the cell-wise standard error of the Poisson intensity, as computed by MCMC.}
\seealso{\link{lgcpPredict}}
\arguments{\item{obj}{an lgcpPredict object}
\item{...}{additional arguments}}
