% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{seintens.lgcpPredict}
\alias{seintens.lgcpPredict}
\title{seintens.lgcpPredict function}
\usage{
\method{seintens}{lgcpPredict}(obj, ...)
}
\arguments{
\item{obj}{an lgcpPredict object}

\item{...}{additional arguments}
}
\value{
the cell-wise standard error of the Poisson intensity, as computed by MCMC.
}
\description{
Accessor function returning the  standard error of the Poisson intensity as an lgcpgrid object.
}
\seealso{
\link{lgcpPredict}
}
