\name{seintens}
\alias{seintens}
\title{seintens function...}
\usage{seintens(obj, ...)}
\description{seintens function}
\details{Generic function to return the standard error of the Poisson Intensity.}
\value{method seintens}
\seealso{\link{lgcpPredict}, \link{seintens.lgcpPredict}}
\arguments{\item{obj}{an object}
\item{...}{additional arguments}}
