\name{selectObsWindow.default}
\alias{selectObsWindow.default}
\title{selectObsWindow...}
\usage{\method{selectObsWindow}{default} (xyt, gridsize, ...)}
\description{selectObsWindow.default function}
\details{Default method, note at present, there is only an implementation for stppp objects.}
\value{currently no default implementation}
\arguments{\item{xyt}{an object}
\item{gridsize}{size of the grid spacing in chosen units (equivalent to the cell width argument in \link{lgcpPredict})}
\item{...}{additional arguments}}
