\name{selectObsWindow}
\alias{selectObsWindow}
\title{selectObsWindow function...}
\usage{selectObsWindow(xyt, ...)}
\description{selectObsWindow function}
\details{Generic function for the purpose of selecting an observation window (or more precisely a bounding box) to contain the FFT grid}
\value{method selectObsWindow}
\arguments{\item{xyt}{an object}
\item{...}{additional arguments}}
