\name{serr.lgcpPredict}
\alias{serr.lgcpPredict}
\title{serr...}
\usage{\method{serr}{lgcpPredict} (obj, ...)}
\description{serr.lgcpPredict function}
\details{Accessor function returning the standard error of relative risk as an lgcpgrid object.}
\value{Standard error of the relative risk as computed by MCMC.}
\seealso{\link{lgcpPredict}}
\arguments{\item{obj}{an lgcpPredict object}
\item{...}{additional arguments}}
