\name{setTxtProgressBar2}
\alias{setTxtProgressBar2}
\title{set the progress bar...}
\usage{setTxtProgressBar2(pb, value, title, label)}
\description{set the progress bar}
\details{update a text progress bar. See help(txtProgressBar) for more info.}
\arguments{\item{pb}{text progress bar object}
\item{value}{new value}
\item{title}{ignored}
\item{label}{text for end of progress bar}}
