\name{showGrid.default}
\alias{showGrid.default}
\title{showGrid...}
\usage{\method{showGrid}{default} (x, y, ...)}
\description{showGrid.default function}
\details{Default method for printing a grid to a screen. Arguments are vectors giving the x any y coordinates of the
centroids.}
\value{plots grid centroids on the current graphics device}
\seealso{\link{showGrid.lgcpPredict}, \link{showGrid.stppp}}
\arguments{\item{x}{an vector of grid values for the x coordinates}
\item{y}{an vector of grid values for the y coordinates}
\item{...}{additional arguments passed to points}}
