% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{showGrid.lgcpPredict}
\alias{showGrid.lgcpPredict}
\title{showGrid.lgcpPredict function}
\usage{
\method{showGrid}{lgcpPredict}(x, ...)
}
\arguments{
\item{x}{an object of class lgcpPredict}

\item{...}{additional arguments  passed to points}
}
\value{
plots grid centroids on the current graphics device
}
\description{
This function displays the FFT grid used on a plot of an \code{lgcpPredict} object.
First plot the object using for example \code{plot(lg)}, where \code{lg} is an object
of class \code{lgcpPredict}, then for any of the plots produced, a call to
\code{showGrid(lg,pch=="+",cex=0.5)} will display the centroids of the FFT grid.
}
\seealso{
\link{lgcpPredict}, \link{showGrid.default}, \link{showGrid.stppp}
}
