\name{showGrid}
\alias{showGrid}
\title{showGrid function}
\usage{
  showGrid(x, ...)
}
\arguments{
  \item{x}{an object}

  \item{...}{additional arguments}
}
\value{
  generic function returning method showGrid
}
\description{
  Generic method for displaying the FFT grid used in
  computation.
}
\seealso{
  \link{showGrid.default}, \link{showGrid.lgcpPredict},
  \link{showGrid.stppp}
}

