\name{smultiply.list}
\alias{smultiply.list}
\title{smultiply...}
\usage{smultiply.list(list, const)}
\description{smultiply.list function}
\details{This function multiplies each element of a list by a scalar constant.}
\value{a list with ith entry the scalar multiple of const * list[[i]]}
\arguments{\item{list}{a list of objects that could be summed using "+"}
\item{const}{a numeric constant}}
