\name{sparsebase}
\alias{sparsebase}
\title{sparsebase function...}
\usage{sparsebase(base)}
\description{sparsebase function}
\details{A function that returns the full precision matrix in sparse format from the base of a block circulant matrix, see ?Matrix::sparseMatrix}
\value{...}
\arguments{\item{base}{base matrix of a block circulant matrix}}
