% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAtRiskClassDef.R
\name{spatialAtRisk.default}
\alias{spatialAtRisk.default}
\title{spatialAtRisk.default function}
\usage{
\method{spatialAtRisk}{default}(X, ...)
}
\arguments{
\item{X}{an object}

\item{...}{additional arguments}
}
\value{
object of class spatialAtRisk
\enumerate{
    \item Brix A, Diggle PJ (2001). Spatiotemporal Prediction for log-Gaussian Cox processes. Journal of the Royal Statistical Society, Series B, 63(4), 823-841.
    \item Diggle P, Rowlingson B, Su T (2005). Point Process Methodology for On-line Spatio-temporal Disease Surveillance. Environmetrics, 16(5), 423-434.
}
}
\description{
The default method for creating a spatialAtRisk object, which attempts to extract x, y and Zm values from the object using \code{xvals},
\code{yvals} and \code{zvals}.
}
\seealso{
\link{lgcpPredict}, link{lgcpSim}, \link{spatialAtRisk.fromXYZ}, \link{spatialAtRisk.im}, \link{spatialAtRisk.function}, \link{spatialAtRisk.SpatialGridDataFrame}, \link{spatialAtRisk.SpatialPolygonsDataFrame}, \link{spatialAtRisk.bivden}, \code{xvals}, \code{yvals}, \code{zvals}
}
