% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAtRiskClassDef.R
\name{spatialAtRisk.im}
\alias{spatialAtRisk.im}
\title{spatialAtRisk.im function}
\usage{
\method{spatialAtRisk}{im}(X, ...)
}
\arguments{
\item{X}{object of class im}

\item{...}{additional arguments}
}
\value{
object of class spatialAtRisk
\enumerate{
    \item Brix A, Diggle PJ (2001). Spatiotemporal Prediction for log-Gaussian Cox processes. Journal of the Royal Statistical Society, Series B, 63(4), 823-841.
    \item Diggle P, Rowlingson B, Su T (2005). Point Process Methodology for On-line Spatio-temporal Disease Surveillance. Environmetrics, 16(5), 423-434.
}
}
\description{
Creates a spatialAtRisk object from a spatstat pixel image (im) object.
}
\seealso{
\link{lgcpPredict}, link{lgcpSim}, \link{spatialAtRisk.default}, \link{spatialAtRisk.fromXYZ}, \link{spatialAtRisk.function}, \link{spatialAtRisk.SpatialGridDataFrame}, \link{spatialAtRisk.SpatialPolygonsDataFrame}, \link{spatialAtRisk.bivden}
}
