\name{spatialIntensities.fromSPDF}
\alias{spatialIntensities.fromSPDF}
\title{spatialIntensities...}
\usage{\method{spatialIntensities}{fromSPDF} (X, xyt, ...)}
\description{spatialIntensities.fromSPDF function}
\details{Extract the spatial intensities from an object of class fromSPDF (as would have been created by spatialAtRisk.SpatialPolygonsDataFrame for example).}
\value{normalised spatial intensities}
\seealso{\link{spatialIntensities}, \link{spatialIntensities.fromXYZ}}
\arguments{\item{X}{an object of class fromSPDF}
\item{xyt}{object of class stppp or a list object of numeric vectors with names $x, $y}
\item{...}{additional arguments}}
