\name{spatialIntensities.fromXYZ}
\alias{spatialIntensities.fromXYZ}
\title{spatialIntensities...}
\usage{\method{spatialIntensities}{fromXYZ} (X, xyt, ...)}
\description{spatialIntensities.fromXYZ function}
\details{Extract the spatial intensities from an object of class fromXYZ (as would have been created by spatialAtRisk for example).}
\value{normalised spatial intensities}
\seealso{\link{spatialIntensities}, \link{spatialIntensities.fromSPDF}}
\arguments{\item{X}{object of class fromXYZ}
\item{xyt}{object of class stppp or a list object of numeric vectors with names $x, $y}
\item{...}{additional arguments}}
