% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAtRiskClassDef.R
\name{spatialIntensities.fromSPDF}
\alias{spatialIntensities.fromSPDF}
\title{spatialIntensities.fromSPDF function}
\usage{
\method{spatialIntensities}{fromSPDF}(X, xyt, ...)
}
\arguments{
\item{X}{an object of class fromSPDF}

\item{xyt}{object of class stppp or a list object of numeric vectors with names $x, $y}

\item{...}{additional arguments}
}
\value{
normalised spatial intensities
}
\description{
Extract the spatial intensities from an object of class fromSPDF (as would have been created by spatialAtRisk.SpatialPolygonsDataFrame for example).
}
\seealso{
\link{spatialIntensities}, \link{spatialIntensities.fromXYZ}
}
