% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAtRiskClassDef.R
\name{spatialIntensities.fromXYZ}
\alias{spatialIntensities.fromXYZ}
\title{spatialIntensities.fromXYZ function}
\usage{
\method{spatialIntensities}{fromXYZ}(X, xyt, ...)
}
\arguments{
\item{X}{object of class fromXYZ}

\item{xyt}{object of class stppp or a list object of numeric vectors with names $x, $y}

\item{...}{additional arguments}
}
\value{
normalised spatial intensities
}
\description{
Extract the spatial intensities from an object of class fromXYZ (as would have been created by spatialAtRisk for example).
}
\seealso{
\link{spatialIntensities}, \link{spatialIntensities.fromSPDF}
}
