\name{spatialIntensities}
\alias{spatialIntensities}
\title{spatialIntensities function...}
\usage{spatialIntensities(X, ...)}
\description{spatialIntensities function}
\details{Generic method for extracting spatial intensities.}
\value{method spatialintensities}
\seealso{\link{spatialIntensities.fromXYZ}, \link{spatialIntensities.fromSPDF}}
\arguments{\item{X}{an object}
\item{...}{additional arguments}}
