\name{SpatialPolygonsDataFrame.stapp}
\alias{SpatialPolygonsDataFrame.stapp}
\title{SpatialPolygonsDataFrame...}
\usage{SpatialPolygonsDataFrame.stapp(from)}
\description{SpatialPolygonsDataFrame.stapp function}
\details{A function to return the SpatialPolygonsDataFrame part of an stapp object}
\value{an object of class SpatialPolygonsDataFrame}
\arguments{\item{from}{stapp object}}
