% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arealMethods.R
\name{SpatialPolygonsDataFrame.stapp}
\alias{SpatialPolygonsDataFrame.stapp}
\title{SpatialPolygonsDataFrame.stapp function}
\usage{
SpatialPolygonsDataFrame.stapp(from)
}
\arguments{
\item{from}{stapp object}
}
\value{
an object of class SpatialPolygonsDataFrame
}
\description{
A function to return the SpatialPolygonsDataFrame part of an stapp object
}
