% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arealMethods.R
\name{stapp.list}
\alias{stapp.list}
\title{stapp.list function}
\usage{
\method{stapp}{list}(obj, ...)
}
\arguments{
\item{obj}{an list object as described above, see ?stapp.SpatialPolygonsDataFrame for further details on the requirements of the list}

\item{...}{additional arguments}
}
\value{
an object of class stapp
}
\description{
A wrapper function for stapp.SpatialPolygonsDataFrame
}
\details{
Construct a space-time aggregated point-process (stapp) object from a list object. The first element of the list 
should be a SpatialPolygonsDataFrame, the second element of the list a counts matrix, the third element of the list a
vector of times, the fourth element a vector giving the bounds of the temporal observation window and the fifth element
a spatstat owin object giving the spatial observation window.
}
