% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arealMethods.R
\name{stapp}
\alias{stapp}
\title{stapp function}
\usage{
stapp(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{additional arguments}
}
\value{
method stapp
}
\description{
Generic function for space-time aggregated point-process data
}
