% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arealMethods.R
\name{stapp.SpatialPolygonsDataFrame}
\alias{stapp.SpatialPolygonsDataFrame}
\title{stapp.SpatialPolygonsDataFrame function}
\usage{
\method{stapp}{SpatialPolygonsDataFrame}(obj, counts, t, tlim, window, ...)
}
\arguments{
\item{obj}{an SpatialPolygonsDataFrame object}

\item{counts}{a (length(t) by N) matrix containing aggregated case counts for each of the geographical regions defined by the SpatialPolygonsDataFrame, where N is the number of regions}

\item{t}{vector of times, for each element of t there should correspond a column in the matrix 'counts'}

\item{tlim}{vector giving the upper and lower bounds of the temporal observation window}

\item{window}{the observation window, of class owin, see ?owin}

\item{...}{additional arguments}
}
\value{
an object of class stapp
}
\description{
Construct a space-time aggregated point-process (stapp) object from a SpatialPolygonsDataFrame (along with some other info)
}
