\name{stppp.list}
\alias{stppp.list}
\title{stppp...}
\usage{\method{stppp}{list} (P, ...)}
\description{stppp.list function}
\details{Construct a space-time planar point pattern from a list object}
\value{an object of class stppp}
\seealso{\link{stppp}, \link{stppp.ppp},}
\arguments{\item{P}{list object containing $data, an (n x 3) matrix corresponding to (x,y,t) values; $tlim, a vector of length 2 givign the observation time window; and $window giving an owin spatial observation winow, see ?owin for more details}
\item{...}{additional arguments}}
