\name{stppp.ppp}
\alias{stppp.ppp}
\title{stppp...}
\usage{\method{stppp}{ppp} (P, t, tlim, ...)}
\description{stppp.ppp function}
\details{Construct a space-time planar point pattern from a ppp object}
\value{an object of class stppp}
\seealso{\link{stppp}, \link{stppp.list}}
\arguments{\item{P}{a spatstat ppp object}
\item{t}{a vector of length P$n}
\item{tlim}{a vector of length 2 specifying the observation time window}
\item{...}{additional arguments}}
