% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stpppClassDef.R
\name{stppp}
\alias{stppp}
\title{stppp function}
\usage{
stppp(P, ...)
}
\arguments{
\item{P}{an object}

\item{...}{additional arguments}
}
\value{
method stppp
}
\description{
Generic function used in the construction of space-time planar point patterns. An stppp object is like a ppp object,
but with extra components for (1) a vector giving the time at whcih the event occurred and (2) a time-window over which observations occurred.
Observations are assumed to occur in the plane and the observation window is assumed not to change over time.
}
\seealso{
\link{stppp}, \link{stppp.ppp}, \link{stppp.list}
}
