\name{summary.lgcpgrid}
\alias{summary.lgcpgrid}
\title{summary...}
\usage{\method{summary}{lgcpgrid} (object, ...)}
\description{summary.lgcpgrid function}
\details{Summary method for lgcp objects. This just applies the summary function to each
of the elements of object$grid.}
\value{Summary per grid, see ?summary for further options}
\seealso{\link{lgcpgrid.list}, \link{lgcpgrid.array}, \link{as.list.lgcpgrid}, \link{print.lgcpgrid}, 
\link{quantile.lgcpgrid}, \link{image.lgcpgrid}, \link{plot.lgcpgrid}}
\arguments{\item{object}{an object of class lgcpgrid}
\item{...}{other arguments}}
