% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{summary.lgcpgrid}
\alias{summary.lgcpgrid}
\title{summary.lgcpgrid function}
\usage{
\method{summary}{lgcpgrid}(object, ...)
}
\arguments{
\item{object}{an object of class lgcpgrid}

\item{...}{other arguments}
}
\value{
Summary per grid, see ?summary for further options
}
\description{
Summary method for lgcp objects. This just applies the summary function to each
of the elements of object$grid.
}
\seealso{
\link{lgcpgrid.list}, \link{lgcpgrid.array}, \link{as.list.lgcpgrid}, \link{print.lgcpgrid}, 
\link{quantile.lgcpgrid}, \link{image.lgcpgrid}, \link{plot.lgcpgrid}
}
