\name{summary.mcmc}
\alias{summary.mcmc}
\title{summary of an mcmc iterator...}
\usage{summary.mcmc(object, ...)}
\description{summary of an mcmc iterator}
\details{print out values of an iterator and reset it. DONT call this
in a loop that uses this iterator - it will reset it. And break.}
\arguments{\item{object}{an mcmc iterator}
\item{...}{other args}}
