% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregateSpatialOnlyPlusParameters.R
\name{target.and.grad.AggregateSpatialPlusPars}
\alias{target.and.grad.AggregateSpatialPlusPars}
\title{target.and.grad.AggregateSpatialPlusPars function}
\usage{
target.and.grad.AggregateSpatialPlusPars(
  GP,
  prior,
  Z,
  Zt,
  eta,
  beta,
  nis,
  cellarea,
  spatial,
  gradtrunc
)
}
\arguments{
\item{GP}{an object constructed using GPrealisation}

\item{prior}{the prior, created using lgcpPrior}

\item{Z}{the design matrix on the full FFT grid}

\item{Zt}{the transpose of the design matrix}

\item{eta}{the model parameter, eta}

\item{beta}{the model parameters, beta}

\item{nis}{cell counts on the FFT grid}

\item{cellarea}{the cell area}

\item{spatial}{the poisson offset}

\item{gradtrunc}{the gradient truncation parameter}
}
\value{
the target and gradient
}
\description{
A function to compute the target and gradient for the Bayesian aggregated point process model. Not for general use.
}
