% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultitypeSpatialOnlyPlusParameters.R
\name{target.and.grad.MultitypespatialPlusPars}
\alias{target.and.grad.MultitypespatialPlusPars}
\title{target.and.grad.MultitypespatialPlusPars function}
\usage{
target.and.grad.MultitypespatialPlusPars(
  GPlist,
  priorlist,
  Zlist,
  Ztlist,
  eta,
  beta,
  nis,
  cellarea,
  spatial,
  gradtrunc
)
}
\arguments{
\item{GPlist}{list of Gaussian processes}

\item{priorlist}{list of priors}

\item{Zlist}{list of design matrices on the FFT gridd}

\item{Ztlist}{list of transposed design matrices}

\item{eta}{LGCP model parameter eta}

\item{beta}{LGCP model parameter beta}

\item{nis}{matrix of cell counts on the extended grid}

\item{cellarea}{the cell area}

\item{spatial}{the poisson offset interpolated onto the correcy grid}

\item{gradtrunc}{gradient truncation paramter}
}
\value{
the target and gradient
}
\description{
A function to compute the taget an gradient for the Bayesian multivariate lgcp
}
