% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialOnlyPlusParameters.R
\name{target.and.grad.spatialPlusPars}
\alias{target.and.grad.spatialPlusPars}
\title{target.and.grad.spatialPlusPars function}
\usage{
target.and.grad.spatialPlusPars(
  GP,
  prior,
  Z,
  Zt,
  eta,
  beta,
  nis,
  cellarea,
  spatial,
  gradtrunc
)
}
\arguments{
\item{GP}{an object created using GPrealisation}

\item{prior}{the model priors, created using lgcpPrior}

\item{Z}{the design matrix on the FFT grid}

\item{Zt}{transpose of the design matrix}

\item{eta}{the paramters, eta}

\item{beta}{the parameters, beta}

\item{nis}{cell counts on the FFT grid}

\item{cellarea}{the cell area}

\item{spatial}{poisson offset}

\item{gradtrunc}{the gradient truncation parameter}
}
\value{
the target and graient for this model
}
\description{
A function to compute the target and gradient for the Bayesian spatial LGCP
}
