% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatioTemporalPlusParameters.R
\name{target.and.grad.SpatioTemporalPlusPars}
\alias{target.and.grad.SpatioTemporalPlusPars}
\title{target.and.grad.SpatioTemporalPlusPars function}
\usage{
target.and.grad.SpatioTemporalPlusPars(
  GP,
  prior,
  Z,
  Zt,
  eta,
  beta,
  nis,
  cellarea,
  spatial,
  gradtrunc,
  ETA0,
  tdiff
)
}
\arguments{
\item{GP}{an object created using the stGPrealisation function}

\item{prior}{the priors for hte model, created using lgcpPrior}

\item{Z}{the design matrix on the FFT grid}

\item{Zt}{the transpose of the design matrix}

\item{eta}{the paramers eta}

\item{beta}{the parameters beta}

\item{nis}{the cell counts on the FFT grid}

\item{cellarea}{the cell area}

\item{spatial}{the poisson offset}

\item{gradtrunc}{the gradient truncation parameter}

\item{ETA0}{the initial value of eta}

\item{tdiff}{vector of time differences between time points}
}
\value{
the target and gradient for the spatiotemporal model.
}
\description{
A function to compute the target and gradient for the Bayesian spatiotemporal LGCP.
}
