\name{toral.cov.mat}
\alias{toral.cov.mat}
\title{toral...}
\usage{toral.cov.mat(xg, yg, sigma, phi, model, additionalparameters)}
\description{toral.cov.mat function}
\details{A function to compute the covariance matrix of a stationary process on a torus.}
\value{circulant covariacne matrix}
\arguments{\item{xg}{x grid}
\item{yg}{y grid}
\item{sigma}{spatial variability parameter}
\item{phi}{spatial decay parameter}
\item{model}{model for covariance, see ?CovarianceFct}
\item{additionalparameters}{additional parameters for covariance structure}}
