% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{touchingowin}
\alias{touchingowin}
\title{touchingowin function}
\usage{
touchingowin(x, y, w)
}
\arguments{
\item{x}{grid centroids in x-direction note this will be expanded into a GRID of (x,y) values in the function}

\item{y}{grid centroids in y-direction note this will be expanded into a GRID of (x,y) values in the function}

\item{w}{an owin or SpatialPolygons object}
}
\value{
vector of TRUE or FALSE according to whether the cell
}
\description{
A function to compute which cells are touching an owin or spatial polygons object
}
