% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{transblack}
\alias{transblack}
\title{transblack function}
\usage{
transblack(alpha = 0.1)
}
\arguments{
\item{alpha}{transparency parameter, see ?rgb}
}
\value{
character string of colour
}
\description{
A function to return a transparent black colour.
}
