% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{transblue}
\alias{transblue}
\title{transblue function}
\usage{
transblue(alpha = 0.1)
}
\arguments{
\item{alpha}{transparency parameter, see ?rgb}
}
\value{
character string of colour
}
\description{
A function to return a transparent blue colour.
}
