% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{transgreen}
\alias{transgreen}
\title{transgreen function}
\usage{
transgreen(alpha = 0.1)
}
\arguments{
\item{alpha}{transparency parameter, see ?rgb}
}
\value{
character string of colour
}
\description{
A function to return a transparent green colour.
}
