\name{updateAMCMC.constanth}
\alias{updateAMCMC.constanth}
\title{updateAMCMC...}
\usage{\method{updateAMCMC}{constanth} (obj, ...)}
\description{updateAMCMC.constanth function}
\details{Updates the \link{constanth} adaptive scheme.}
\value{update and return current h for scheme}
\seealso{\link{constanth}}
\arguments{\item{obj}{an object}
\item{...}{additional arguments}}
