\name{updateAMCMC.andrieuthomsh}
\alias{updateAMCMC.andrieuthomsh}
\title{updateAMCMC...}
\usage{\method{updateAMCMC}{andrieuthomsh} (obj, ...)}
\description{updateAMCMC.andrieuthomsh function}
\details{Updates the \link{andrieuthomsh} adaptive scheme.}
\value{update and return current h for scheme}
\references{\enumerate{
\item Andrieu C, Thoms J (2008). A tutorial on adaptive MCMC. Statistics and Computing, 18(4), 343-373.
\item Robbins H, Munro S (1951). A Stochastic Approximation Methods. The Annals of Mathematical Statistics, 22(3), 400-407.
\item Roberts G, Rosenthal J (2001). Optimal Scaling for Various Metropolis-Hastings Algorithms. Statistical Science, 16(4), 351-367.
}}
\seealso{\link{andrieuthomsh}}
\arguments{\item{obj}{an object}
\item{...}{additional arguments}}
