% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptiveSchemes.R
\name{updateAMCMC.constanth}
\alias{updateAMCMC.constanth}
\title{updateAMCMC.constanth function}
\usage{
\method{updateAMCMC}{constanth}(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{additional arguments}
}
\value{
update and return current h for scheme
}
\description{
Updates the \link{constanth} adaptive scheme.
}
\seealso{
\link{constanth}
}
