\name{varfield.lgcpPredictINLA}
\alias{varfield.lgcpPredictINLA}
\title{varfield...}
\usage{\method{varfield}{lgcpPredictINLA} (obj, ...)}
\description{varfield.lgcpPredictINLA function}
\details{A function to return the variance of the latent field from a call to lgcpPredictINLA output.}
\value{the variance of the latent field}
\arguments{\item{obj}{an object of class lgcpPredictINLA}
\item{...}{other arguments}}
