\name{varfield.lgcpPredict}
\alias{varfield.lgcpPredict}
\title{varfield...}
\usage{\method{varfield}{lgcpPredict} (obj, ...)}
\description{varfield.lgcpPredict function}
\details{This is an accessor function for objects of class \code{lgcpPredict} and returns the variance of the 
field Y as an lgcpgrid object.}
\value{returns the cell-wise variance of Y computed via Monte Carlo.}
\seealso{\link{lgcpPredict}}
\arguments{\item{obj}{an object of class lgcpPredict}
\item{...}{additional arguments}}
