\name{xvals.default}
\alias{xvals.default}
\title{xvals...}
\usage{\method{xvals}{default} (obj, ...)}
\description{xvals.default function}
\details{Default method for extracting 'x values' looks for $X, $x in that order.}
\value{the x values}
\seealso{\link{xvals},  \link{yvals}, \link{zvals}, \link{yvals.default}, \link{zvals.default}, \link{xvals.fromXYZ}, \link{yvals.fromXYZ}, \link{zvals.fromXYZ}, \link{xvals.SpatialGridDataFrame}, \link{yvals.SpatialGridDataFrame}, \link{zvals.SpatialGridDataFrame}}
\arguments{\item{obj}{an object}
\item{...}{additional arguments}}
