\name{xvals.fromXYZ}
\alias{xvals.fromXYZ}
\title{xvals...}
\usage{\method{xvals}{fromXYZ} (obj, ...)}
\description{xvals.fromXYZ function}
\details{Method for extracting 'x values' from an object of class fromXYZ}
\value{the x values}
\seealso{\link{xvals},  \link{yvals}, \link{zvals}, \link{xvals.default}, \link{yvals.default}, \link{zvals.default}, \link{yvals.fromXYZ}, \link{zvals.fromXYZ}, \link{xvals.SpatialGridDataFrame}, \link{yvals.SpatialGridDataFrame}, \link{zvals.SpatialGridDataFrame}}
\arguments{\item{obj}{a spatialAtRisk object}
\item{...}{additional arguments}}
