\name{xvals.lgcpPredict}
\alias{xvals.lgcpPredict}
\title{xvals...}
\usage{\method{xvals}{lgcpPredict} (obj, ...)}
\description{xvals.lgcpPredict function}
\details{Gets the x-coordinates of the centroids of the prediction grid.}
\value{the x coordinates of the centroids of the grid}
\seealso{\link{lgcpPredict}}
\arguments{\item{obj}{an object of class lgcpPredict}
\item{...}{additional arguments}}
