\name{xvals}
\alias{xvals}
\title{xvals function...}
\usage{xvals(obj, ...)}
\description{xvals function}
\details{Generic for extractign the 'x values' from an object.}
\value{the xvals method}
\seealso{\link{yvals}, \link{zvals}, \link{xvals.default}, \link{yvals.default}, \link{zvals.default}, \link{xvals.fromXYZ}, \link{yvals.fromXYZ}, \link{zvals.fromXYZ}, \link{xvals.SpatialGridDataFrame}, \link{yvals.SpatialGridDataFrame}, \link{zvals.SpatialGridDataFrame}}
\arguments{\item{obj}{an object of class spatialAtRisk}
\item{...}{additional arguments}}
