\name{xvals.SpatialGridDataFrame}
\alias{xvals.SpatialGridDataFrame}
\title{xvals...}
\usage{\method{xvals}{SpatialGridDataFrame} (obj, ...)}
\description{xvals.SpatialGridDataFrame function}
\details{Method for extracting 'x values' from an object of class spatialGridDataFrame}
\value{the x values}
\seealso{\link{xvals},  \link{yvals}, \link{zvals}, \link{xvals.default}, \link{yvals.default}, \link{zvals.default}, \link{xvals.fromXYZ}, \link{yvals.fromXYZ}, \link{zvals.fromXYZ}, \link{yvals.SpatialGridDataFrame}, \link{zvals.SpatialGridDataFrame}}
\arguments{\item{obj}{an object}
\item{...}{additional arguments}}
