\name{yvals.default}
\alias{yvals.default}
\title{yvals...}
\usage{\method{yvals}{default} (obj, ...)}
\description{yvals.default function}
\details{Default method for extracting 'y values' looks for $Y, $y in that order.}
\value{the y values}
\seealso{\link{xvals},  \link{yvals}, \link{zvals}, \link{xvals.default}, \link{zvals.default}, \link{xvals.fromXYZ}, \link{yvals.fromXYZ}, \link{zvals.fromXYZ}, \link{xvals.SpatialGridDataFrame}, \link{yvals.SpatialGridDataFrame}, \link{zvals.SpatialGridDataFrame}}
\arguments{\item{obj}{an object}
\item{...}{additional arguments}}
