\name{yvals.fromXYZ}
\alias{yvals.fromXYZ}
\title{yvals...}
\usage{\method{yvals}{fromXYZ} (obj, ...)}
\description{yvals.fromXYZ function}
\details{Method for extracting 'y values' from an object of class fromXYZ}
\value{the y values}
\seealso{\link{xvals},  \link{yvals}, \link{zvals}, \link{xvals.default}, \link{yvals.default}, \link{zvals.default}, \link{xvals.fromXYZ}, \link{zvals.fromXYZ}, \link{xvals.SpatialGridDataFrame}, \link{yvals.SpatialGridDataFrame}, \link{zvals.SpatialGridDataFrame}}
\arguments{\item{obj}{a spatialAtRisk object}
\item{...}{additional arguments}}
