\name{yvals.lgcpPredict}
\alias{yvals.lgcpPredict}
\title{yvals...}
\usage{\method{yvals}{lgcpPredict} (obj, ...)}
\description{yvals.lgcpPredict function}
\details{Gets the y-coordinates of the centroids of the prediction grid.}
\value{the y coordinates of the centroids of the grid}
\seealso{\link{lgcpPredict}}
\arguments{\item{obj}{an object of class lgcpPredict}
\item{...}{additional arguments}}
