\name{yvals.SpatialGridDataFrame}
\alias{yvals.SpatialGridDataFrame}
\title{yvals...}
\usage{\method{yvals}{SpatialGridDataFrame} (obj, ...)}
\description{yvals.SpatialGridDataFrame function}
\details{Method for extracting 'y values' from an object of class SpatialGridDataFrame}
\value{the y values}
\seealso{\link{xvals},  \link{yvals}, \link{zvals}, \link{xvals.default}, \link{yvals.default}, \link{zvals.default}, \link{xvals.fromXYZ}, \link{yvals.fromXYZ}, \link{zvals.fromXYZ}, \link{xvals.SpatialGridDataFrame}, \link{zvals.SpatialGridDataFrame}}
\arguments{\item{obj}{an object}
\item{...}{additional arguments}}
