% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAtRiskClassDef.R
\name{yvals.fromXYZ}
\alias{yvals.fromXYZ}
\title{yvals.fromXYZ function}
\usage{
\method{yvals}{fromXYZ}(obj, ...)
}
\arguments{
\item{obj}{a spatialAtRisk object}

\item{...}{additional arguments}
}
\value{
the y values
}
\description{
Method for extracting 'y values' from an object of class fromXYZ
}
\seealso{
\link{xvals},  \link{yvals}, \link{zvals}, \link{xvals.default}, \link{yvals.default}, \link{zvals.default}, \link{xvals.fromXYZ}, \link{zvals.fromXYZ}, \link{xvals.SpatialGridDataFrame}, \link{yvals.SpatialGridDataFrame}, \link{zvals.SpatialGridDataFrame}
}
