% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{yvals.lgcpPredict}
\alias{yvals.lgcpPredict}
\title{yvals.lgcpPredict function}
\usage{
\method{yvals}{lgcpPredict}(obj, ...)
}
\arguments{
\item{obj}{an object of class lgcpPredict}

\item{...}{additional arguments}
}
\value{
the y coordinates of the centroids of the grid
}
\description{
Gets the y-coordinates of the centroids of the prediction grid.
}
\seealso{
\link{lgcpPredict}
}
