\name{yvals}
\alias{yvals}
\title{yvals function...}
\usage{yvals(obj, ...)}
\description{yvals function}
\details{Generic for extractign the 'y values' from an object.}
\value{the yvals method}
\seealso{\link{xvals}, \link{zvals}, \link{xvals.default}, \link{yvals.default}, \link{zvals.default}, \link{xvals.fromXYZ}, \link{yvals.fromXYZ}, \link{zvals.fromXYZ}, \link{xvals.SpatialGridDataFrame}, \link{yvals.SpatialGridDataFrame}, \link{zvals.SpatialGridDataFrame}}
\arguments{\item{obj}{an object of class spatialAtRisk}
\item{...}{additional arguments}}
