\name{zvals.fromXYZ}
\alias{zvals.fromXYZ}
\title{zvals...}
\usage{\method{zvals}{fromXYZ} (obj, ...)}
\description{zvals.fromXYZ function}
\details{Method for extracting 'z values' from an object of class fromXYZ}
\value{the z values}
\seealso{\link{xvals},  \link{yvals}, \link{zvals}, \link{xvals.default}, \link{yvals.default}, \link{zvals.default}, \link{xvals.fromXYZ}, \link{yvals.fromXYZ}, \link{xvals.SpatialGridDataFrame}, \link{yvals.SpatialGridDataFrame}, \link{zvals.SpatialGridDataFrame}}
\arguments{\item{obj}{a spatialAtRisk object}
\item{...}{additional arguments}}
