\name{zvals.default}
\alias{zvals.default}
\title{zvals...}
\usage{\method{zvals}{default} (obj, ...)}
\description{zvals.default function}
\details{Default method for extracting 'z values' looks for $Zm, $Z, $z in that order.}
\value{the x values}
\seealso{\link{xvals},  \link{yvals}, \link{zvals}, \link{xvals.default}, \link{yvals.default}, \link{xvals.fromXYZ}, \link{yvals.fromXYZ}, \link{zvals.fromXYZ}, \link{xvals.SpatialGridDataFrame}, \link{yvals.SpatialGridDataFrame}, \link{zvals.SpatialGridDataFrame}}
\arguments{\item{obj}{an object}
\item{...}{additional arguments}}
