\name{zvals}
\alias{zvals}
\title{zvals function...}
\usage{zvals(obj, ...)}
\description{zvals function}
\details{Generic for extractign the 'z values' from an object.}
\value{the zvals method}
\seealso{\link{xvals},  \link{yvals}, \link{xvals.default}, \link{yvals.default}, \link{zvals.default}, \link{xvals.fromXYZ}, \link{yvals.fromXYZ}, \link{zvals.fromXYZ}, \link{xvals.SpatialGridDataFrame}, \link{yvals.SpatialGridDataFrame}, \link{zvals.SpatialGridDataFrame}}
\arguments{\item{obj}{an object}
\item{...}{additional arguments}}
