\name{zvals.SpatialGridDataFrame}
\alias{zvals.SpatialGridDataFrame}
\title{zvals...}
\usage{\method{zvals}{SpatialGridDataFrame} (obj, ...)}
\description{zvals.SpatialGridDataFrame function}
\details{Method for extracting 'z values' from an object of class SpatialGridDataFrame}
\value{the z values}
\seealso{\link{xvals},  \link{yvals}, \link{zvals}, \link{xvals.default}, \link{yvals.default}, \link{zvals.default}, \link{xvals.fromXYZ}, \link{yvals.fromXYZ}, \link{zvals.fromXYZ}, \link{xvals.SpatialGridDataFrame}, \link{yvals.SpatialGridDataFrame}}
\arguments{\item{obj}{an object}
\item{...}{additional arguments}}
