% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectObsWindow.R
\name{affine.fromSPDF}
\alias{affine.fromSPDF}
\title{affine.fromSPDF function}
\usage{
\method{affine}{fromSPDF}(X, mat, ...)
}
\arguments{
\item{X}{an object of class fromSPDF}

\item{mat}{matrix of affine transformation}

\item{...}{additional arguments}
}
\value{
the object acted on by the transformation matrix
}
\description{
An affine transformation of an object of class \code{fromSPDF}
}
