% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometryHandling.R
\name{aggCovInfo.ArealWeightedMean}
\alias{aggCovInfo.ArealWeightedMean}
\title{aggCovInfo.ArealWeightedMean function}
\usage{
\method{aggCovInfo}{ArealWeightedMean}(obj, regwts, ...)
}
\arguments{
\item{obj}{an ArealWeightedMean object}

\item{regwts}{regional (areal) weighting vector}

\item{...}{additional arguments}
}
\value{
Areal weighted mean.
}
\description{
Aggregation via weighted mean.
}
