% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAtRiskClassDef.R
\name{as.fromXYZ}
\alias{as.fromXYZ}
\title{as.fromXYZ function}
\usage{
as.fromXYZ(X, ...)
}
\arguments{
\item{X}{an object}

\item{...}{additional arguments}
}
\value{
generic function returning method as.fromXYZ
}
\description{
Generic function for conversion to a fromXYZ object (eg as would have been produced by spatialAtRisk for example.)
}
\seealso{
\link{as.im.fromXYZ}, \link{as.im.fromSPDF}, \link{as.im.fromFunction}, \link{as.fromXYZ}
}
