% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arealMethods.R
\name{as.owin.stapp}
\alias{as.owin.stapp}
\title{as.owin.stapp function}
\usage{
\method{as.owin}{stapp}(W, ..., fatal = TRUE)
}
\arguments{
\item{W}{see ?as.owin}

\item{...}{see ?as.owin}

\item{fatal}{see ?as.owin}
}
\value{
an owin object
}
\description{
A function to extract the SpatialPolygons part of W and return it as an owin object.
}
