% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arealMethods.R
\name{as.owinlist.stapp}
\alias{as.owinlist.stapp}
\title{as.owinlist.stapp function}
\usage{
\method{as.owinlist}{stapp}(obj, dmin = 0, check = TRUE, ...)
}
\arguments{
\item{obj}{an stapp object}

\item{dmin}{purpose is to simplify the SpatialPolygons. A numeric value giving the smallest permissible length of an edge. See ? simplify.owin}

\item{check}{whether or not to use spatstat functions to check the validity of SpatialPolygons objects}

\item{...}{additional arguments}
}
\value{
a list of owin objects corresponding to the constituent Polygons objects
}
\description{
A function to create a list of owin objects from a stapp
}
