% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAtRiskClassDef.R
\name{as.SpatialGridDataFrame.fromXYZ}
\alias{as.SpatialGridDataFrame.fromXYZ}
\title{as.SpatialGridDataFrame.fromXYZ function}
\usage{
\method{as.SpatialGridDataFrame}{fromXYZ}(obj, ...)
}
\arguments{
\item{obj}{an object of class spatialAtRisk}

\item{...}{additional arguments}
}
\value{
an object of class SpatialGridDataFrame
}
\description{
Method for converting objects of class fromXYZ into those of class SpatialGridDataFrame
}
\seealso{
\link{as.SpatialGridDataFrame}
}
