% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAtRiskClassDef.R
\name{as.SpatialGridDataFrame}
\alias{as.SpatialGridDataFrame}
\title{as.SpatialGridDataFrame function}
\usage{
as.SpatialGridDataFrame(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{additional arguments}
}
\value{
method as.SpatialGridDataFrame
}
\description{
Generic method for convertign to an object of class SpatialGridDataFrame.
}
\seealso{
\link{as.SpatialGridDataFrame.fromXYZ}
}
