% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{as.SpatialPixelsDataFrame.lgcpgrid}
\alias{as.SpatialPixelsDataFrame.lgcpgrid}
\title{as.SpatialPixelsDataFrame.lgcpgrid function}
\usage{
\method{as.SpatialPixelsDataFrame}{lgcpgrid}(obj, ...)
}
\arguments{
\item{obj}{an lgcpgrid object}

\item{...}{additional arguments to be passed to SpatialPoints, eg a proj4string}
}
\value{
Either a SpatialPixelsDataFrame, or a list consisting of SpatialPixelsDataFrame objects.
}
\description{
Method to convert lgcpgrid objects to SpatialPixelsDataFrame objects.
}
