% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialOnly.R
\name{circulant.matrix}
\alias{circulant.matrix}
\title{circulant.matrix function}
\usage{
\method{circulant}{matrix}(x, ...)
}
\arguments{
\item{x}{a matrix object}

\item{...}{additional arguments}
}
\value{
If x is a matrix whose columns are the bases of the sub-blocks of a block circulant matrix, then this function returns the block circulant matrix of interest.
}
\description{
If x is a matrix whose columns are the bases of the sub-blocks of a block circulant matrix, then this function returns the 
block circulant matrix of interest.
}
