% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptiveSchemes.R
\name{constanth}
\alias{constanth}
\title{constanth function}
\usage{
constanth(h)
}
\arguments{
\item{h}{an object}
}
\value{
object of class constanth
}
\description{
This function is used to set up a constant acceptance scheme in the argument 
\code{mcmc.control} of the function \code{lgcpPredict}. The scheme is set via
the \code{mcmcpars} function.
}
\examples{
constanth(0.01)
}
\seealso{
\link{mcmcpars}, \link{lgcpPredict}
}
