% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporalAtRiskClassDef.R
\name{constantInTime}
\alias{constantInTime}
\title{constantInTime function}
\usage{
constantInTime(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{additional arguments}
}
\value{
method constantInTime
}
\description{
Generic function for creating constant-in-time temporalAtRisk objects, that is for models where mu(t) can be assumed to
be constant in time. The assumption being that the global at-risk population does not change in size over time.
}
\details{
For further details of temporalAtRisk objects, see ?temporalAtRisk>
}
\seealso{
\link{temporalAtRisk}, \link{spatialAtRisk}, \link{temporalAtRisk.numeric}, \link{temporalAtRisk.function}, \link{constantInTime.numeric}, \link{constantInTime.stppp}, \link{print.temporalAtRisk}, \link{plot.temporalAtRisk}
}
