% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterEstimation.R
\name{Cvb}
\alias{Cvb}
\title{Cvb function}
\usage{
Cvb(xyt, spatial.intensity, N = 100, spatial.covmodel, covpars)
}
\arguments{
\item{xyt}{object of class stppp}

\item{spatial.intensity}{bivariate density estimate of lambda, an object of class im (produced from density.ppp for example)}

\item{N}{number of integration points}

\item{spatial.covmodel}{spatial covariance model}

\item{covpars}{additional covariance parameters}
}
\value{
a function, see below.
Computes Monte carlo estimate of function C(v;beta) in Brix and Diggle 2001 pp 829 (... note later corrigendum to paper (2003) corrects the expression given in this paper)
}
\description{
This function is used in \code{thetaEst} to estimate the temporal correlation parameter, theta.
}
\references{
\enumerate{
    \item Benjamin M. Taylor, Tilman M. Davies, Barry S. Rowlingson, Peter J. Diggle (2013). Journal of Statistical Software, 52(4), 1-40. URL http://www.jstatsoft.org/v52/i04/
    \item Brix A, Diggle PJ (2001). Spatiotemporal Prediction for log-Gaussian Cox processes. Journal of the Royal Statistical Society, Series B, 63(4), 823-841.
}
}
\seealso{
\link{thetaEst}
}
