% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterEstimation.R
\name{density.stppp}
\alias{density.stppp}
\title{density.stppp function}
\usage{
\method{density}{stppp}(x, bandwidth = NULL, ...)
}
\arguments{
\item{x}{an stppp object}

\item{bandwidth}{'bandwidth' parameter, equivanent to parameter sigma in ?density.ppp ie standard deviation of isotropic Gaussian smoothing kernel.}

\item{...}{additional arguments to be passed to density.ppp}
}
\value{
bivariate density estimate of xyt; not this is a wrapper function for density.ppp
}
\description{
A wrapper function for \link{density.ppp}.
}
\seealso{
\link{density.ppp}
}
