\name{discreteWindow}
\alias{discreteWindow}
\title{discreteWindow function...}
\usage{discreteWindow(obj, ...)}
\description{discreteWindow function}
\details{Generic function for extracting the FFT discrete window.}
\value{method discreteWindow}
\seealso{\link{discreteWindow.lgcpPredict}}
\arguments{\item{obj}{an object}
\item{...}{additional arguments}}
