% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{discreteWindow}
\alias{discreteWindow}
\title{discreteWindow function}
\usage{
discreteWindow(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{additional arguments}
}
\value{
method discreteWindow
}
\description{
Generic function for extracting the FFT discrete window.
}
\seealso{
\link{discreteWindow.lgcpPredict}
}
