% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialOnly.R
\name{eigenfrombase}
\alias{eigenfrombase}
\title{eigenfrombase function}
\usage{
eigenfrombase(x)
}
\arguments{
\item{x}{the base matrix}
}
\value{
the eigenvalues
}
\description{
A function to compute the eigenvalues of an SPD block circulant matrix given the base matrix.
}
