% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{expectation}
\alias{expectation}
\title{expectation function}
\usage{
expectation(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{additional arguments}
}
\value{
method expectation
}
\description{
Generic function used in the computation of Monte Carlo expectations.
}
