\name{extract}
\alias{extract}
\title{extract function}
\usage{
extract(obj, ...)
}
\arguments{
  \item{obj}{an object}

  \item{...}{additional arguments}
}
\value{
method extract
}
\description{
Generic function for extracting information dumped to disk.
See \link{extract.lgcpPredict} for further information.
}
\seealso{
\link{extract.lgcpPredict}
}

