% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftgrid.R
\name{fftgrid}
\alias{fftgrid}
\title{fftgrid function}
\usage{
fftgrid(xyt, M, N, spatial, sigma, phi, model, covpars, inclusion = "touching")
}
\arguments{
\item{xyt}{object of class stppp}

\item{M}{number of centroids in x-direction}

\item{N}{number of centroids in y-direction}

\item{spatial}{an object of class spatialAtRisk}

\item{sigma}{scaling paramter for spatial covariance function, see Brix and Diggle (2001)}

\item{phi}{scaling paramter for spatial covariance function, see Brix and Diggle (2001)}

\item{model}{correlation type see ?CovarianceFct}

\item{covpars}{vector of additional parameters for certain classes of covariance function (eg Matern), these must be supplied in the order given in ?CovarianceFct}

\item{inclusion}{criterion for cells being included into observation window. Either 'touching' or 'centroid'. The former includes all cells that touch the observation window, the latter includes all cells whose centroids are inside the observation window.}
}
\value{
fft objects for use in MALA
}
\description{
! As of lgcp version 0.9-5, this function is no longer used !
}
\details{
\bold{Advanced use only.} Computes various quantities for use in \code{lgcpPredict},
\code{lgcpSim} .
}
