% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftgrid.R
\name{fftinterpolate.fromXYZ}
\alias{fftinterpolate.fromXYZ}
\title{interpolate.fromXYZ function}
\usage{
\method{fftinterpolate}{fromXYZ}(spatial, mcens, ncens, ext, ...)
}
\arguments{
\item{spatial}{objects of class spatialAtRisk}

\item{mcens}{x-coordinates of interpolation grid in extended space}

\item{ncens}{y-coordinates of interpolation grid in extended space}

\item{ext}{integer multiple by which grid should be extended, default is 2. Generally this will not need to be altered, but if the spatial correlation decays slowly, increasing 'ext' may be necessary.}

\item{...}{additional arguments}
}
\value{
matrix of interpolated values
}
\description{
This method performs interpolation within the function \code{fftgrid} for \code{fromXYZ} objects.
}
\seealso{
\link{fftgrid}, \link{spatialAtRisk.fromXYZ}
}
