\name{fftinterpolate}
\alias{fftinterpolate}
\title{fftinterpolate function}
\usage{
  fftinterpolate(spatial, ...)
}
\arguments{
  \item{spatial}{an object}

  \item{...}{additional arguments}
}
\value{
  method fftinterpolate
}
\description{
  Generic function used for computing interpolations used
  in the function \link{fftgrid}.
}
\seealso{
  \link{fftgrid}
}

