% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridAverageMethods.R
\name{GAfinalise}
\alias{GAfinalise}
\title{GAfinalise function}
\usage{
GAfinalise(F, ...)
}
\arguments{
\item{F}{an object}

\item{...}{additional arguments}
}
\value{
method GAfinalise
}
\description{
Generic function defining the the finalisation step for the \code{gridAverage} class of functions.
The function is called invisibly within \code{MALAlgcp} and facilitates the computation of
Monte Carlo Averages online.
}
\seealso{
\link{setoutput}, \link{GAinitialise}, \link{GAupdate}, \link{GAreturnvalue}
}
