% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridAverageMethods.R
\name{GAinitialise}
\alias{GAinitialise}
\title{GAinitialise function}
\usage{
GAinitialise(F, ...)
}
\arguments{
\item{F}{an object}

\item{...}{additional arguments}
}
\value{
method GAinitialise
}
\description{
Generic function defining the the initialisation step for the \code{gridAverage} class of functions. 
The function is called invisibly within \code{MALAlgcp} and facilitates the computation of
Monte Carlo Averages online.
}
\seealso{
\link{setoutput}, \link{GAupdate}, \link{GAfinalise}, \link{GAreturnvalue}
}
