\name{GAreturnvalue.MonteCarloAverage}
\alias{GAreturnvalue.MonteCarloAverage}
\title{GAreturnvalue.MonteCarloAverage function}
\usage{
\method{GAreturnvalue}{MonteCarloAverage}(F, ...)
}
\arguments{
  \item{F}{an object of class MonteCarloAverage}

  \item{...}{additional arguments}
}
\value{
results from MonteCarloAverage
}
\description{
Returns the required Monte Carlo average.
}
\seealso{
\link{MonteCarloAverage}, \link{setoutput},
\link{GAinitialise}, \link{GAupdate}, \link{GAfinalise},
\link{GAreturnvalue}
}

