% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridAverageMethods.R
\name{GAreturnvalue}
\alias{GAreturnvalue}
\title{GAreturnvalue function}
\usage{
GAreturnvalue(F, ...)
}
\arguments{
\item{F}{an object}

\item{...}{additional arguments}
}
\value{
method GAreturnvalue
}
\description{
Generic function defining the the returned value for the \code{gridAverage} class of functions.
The function is called invisibly within \code{MALAlgcp} and facilitates the computation of
Monte Carlo Averages online.
}
\seealso{
\link{setoutput}, \link{GAinitialise}, \link{GAupdate}, \link{GAfinalise}
}
