% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridAverageMethods.R
\name{GAupdate}
\alias{GAupdate}
\title{GAupdate function}
\usage{
GAupdate(F, ...)
}
\arguments{
\item{F}{an object}

\item{...}{additional arguments}
}
\value{
method GAupdate
}
\description{
Generic function defining the the update step for the \code{gridAverage} class of functions.
The function is called invisibly within \code{MALAlgcp} and facilitates the computation of
Monte Carlo Averages online.
}
\seealso{
\link{setoutput}, \link{GAinitialise}, \link{GAfinalise}, \link{GAreturnvalue}
}
