% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCounts.R
\name{getCellCounts}
\alias{getCellCounts}
\title{getCellCounts function}
\usage{
getCellCounts(x, y, xgrid, ygrid)
}
\arguments{
\item{x}{x-coordinates of events}

\item{y}{y-coordinates of events}

\item{xgrid}{x-coordinates of grid centroids}

\item{ygrid}{y-coordinates of grid centroids}
}
\value{
The number of observations in each grid cell.
}
\description{
This function is used to count the number of observations falling inside grid cells.
}
