% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{getCovParameters.list}
\alias{getCovParameters.list}
\title{getCovParameters.list function}
\usage{
\method{getCovParameters}{list}(obj, ...)
}
\arguments{
\item{obj}{an list object}

\item{...}{additional arguments}
}
\value{
...
}
\description{
Internal function for retrieving covariance parameters. not indended for general use.
}
