% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectObsWindow.R
\name{getRotation}
\alias{getRotation}
\title{getRotation function}
\usage{
getRotation(xyt, ...)
}
\arguments{
\item{xyt}{an object}

\item{...}{additional arguments}
}
\value{
method getRotation
}
\description{
Generic function for the computation of rotation matrices.
}
\seealso{
\link{getRotation.stppp}
}
