% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridFunctionMethods.R
\name{GFfinalise}
\alias{GFfinalise}
\title{GFfinalise function}
\usage{
GFfinalise(F, ...)
}
\arguments{
\item{F}{an object}

\item{...}{additional arguments}
}
\value{
method GFfinalise
}
\description{
Generic function defining the the finalisation step for the \code{gridFunction} class of objects. 
The function is called invisibly within \code{MALAlgcp} and facilitates the dumping of data to disk
}
\seealso{
\link{setoutput}, \link{GFinitialise}, \link{GFupdate}, \link{GFreturnvalue}
}
