% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridFunctionMethods.R
\name{GFinitialise.dump2dir}
\alias{GFinitialise.dump2dir}
\title{GFinitialise.dump2dir function}
\usage{
\method{GFinitialise}{dump2dir}(F, ...)
}
\arguments{
\item{F}{an object of class dump2dir}

\item{...}{additional arguments}
}
\value{
creates initialisation file and folder
}
\description{
Creates a directory (if necessary) and allocates space for a netCDF dump.
}
\seealso{
\link{dump2dir}, \link{setoutput}, \link{GFinitialise}, \link{GFupdate}, \link{GFfinalise}, \link{GFreturnvalue}
}
