% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridFunctionMethods.R
\name{GFreturnvalue.dump2dir}
\alias{GFreturnvalue.dump2dir}
\title{GFreturnvalue.dump2dir function}
\usage{
\method{GFreturnvalue}{dump2dir}(F, ...)
}
\arguments{
\item{F}{an object}

\item{...}{additional arguments}
}
\value{
display where files have been written to
}
\description{
This function returns the name of the directory the netCDF file was written to.
}
\seealso{
\link{dump2dir}, \link{setoutput}, \link{GFinitialise}, \link{GFupdate}, \link{GFfinalise}, \link{GFreturnvalue}
}
