% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridFunctionMethods.R
\name{GFreturnvalue}
\alias{GFreturnvalue}
\title{GFreturnvalue function}
\usage{
GFreturnvalue(F, ...)
}
\arguments{
\item{F}{an object}

\item{...}{additional arguments}
}
\value{
method GFreturnvalue
}
\description{
Generic function defining the the returned value for the \code{gridFunction} class of objects. 
The function is called invisibly within \code{MALAlgcp} and facilitates the dumping of data to disk
}
\seealso{
\link{setoutput}, \link{GFinitialise}, \link{GFupdate}, \link{GFfinalise}
}
