% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridFunctionMethods.R
\name{GFupdate.dump2dir}
\alias{GFupdate.dump2dir}
\title{GFupdate.dump2dir function}
\usage{
\method{GFupdate}{dump2dir}(F, ...)
}
\arguments{
\item{F}{an object}

\item{...}{additional arguments}
}
\value{
saves latent field
}
\description{
This function gets the required information from \code{MALAlgcp} and writes the data to the netCDF file.
}
\seealso{
\link{dump2dir}, \link{setoutput}, \link{GFinitialise}, \link{GFupdate}, \link{GFfinalise}, \link{GFreturnvalue}
}
