% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridFunctionMethods.R
\name{GFupdate}
\alias{GFupdate}
\title{GFupdate function}
\usage{
GFupdate(F, ...)
}
\arguments{
\item{F}{an object}

\item{...}{additional arguments}
}
\value{
method GFupdate
}
\description{
Generic function defining the the update step for the \code{gridFunction} class of objects. 
The function is called invisibly within \code{MALAlgcp} and facilitates the dumping of data to disk
}
\seealso{
\link{setoutput}, \link{GFinitialise}, \link{GFfinalise}, \link{GFreturnvalue}
}
