% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{gridav}
\alias{gridav}
\title{gridav function}
\usage{
gridav(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{additional arguments}
}
\value{
method gridav
}
\description{
A generic function for returning \code{gridmeans} objects.
}
\seealso{
\link{setoutput}, \link{lgcpgrid}
}
