\name{gridfun}
\alias{gridfun}
\title{gridfun function}
\usage{
gridfun(obj, ...)
}
\arguments{
  \item{obj}{an object}

  \item{...}{additional arguments}
}
\value{
method gridfun
}
\description{
A generic function for returning \code{gridfunction}
objects.
}
\seealso{
\link{setoutput}, \link{lgcpgrid}
}

