% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{gridfun}
\alias{gridfun}
\title{gridfun function}
\usage{
gridfun(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{additional arguments}
}
\value{
method gridfun
}
\description{
A generic function for returning \code{gridfunction} objects.
}
\seealso{
\link{setoutput}, \link{lgcpgrid}
}
