% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{gridInWindow}
\alias{gridInWindow}
\title{gridInWindow function}
\usage{
gridInWindow(xvals, yvals, win, inclusion = "touching")
}
\arguments{
\item{xvals}{x coordinates}

\item{yvals}{y coordinates}

\item{win}{owin object}

\item{inclusion}{criterion for cells being included into observation window. Either 'touching' or 'centroid'. The former includes all cells that touch the observation window, the latter includes all cells whose centroids are inside the observation window.}
}
\value{
matrix of TRUE/FALSE, which elements of the grid are inside the observation window win
}
\description{
For the grid defined by x-coordinates, xvals, and y-coordinates, yvals, and an owin object W, this function just returns
a logical matrix M, whose [i,j] entry is TRUE if the point(xvals[i], yvals[j]) is inside the observation window.
}
