% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpInternals.R
\name{gu}
\alias{gu}
\title{gu function}
\usage{
gu(u, sigma, phi, model, additionalparameters)
}
\arguments{
\item{u}{distance}

\item{sigma}{variance parameter, see Brix and Diggle (2001)}

\item{phi}{scale parameter, see Brix and Diggle (2001)}

\item{model}{correlation type, see ?CovarianceFct}

\item{additionalparameters}{vector of additional parameters for certain classes of covariance function (eg Matern), these must be supplied in the order given in ?CovarianceFct}
}
\value{
this is just a wrapper for CovarianceFct
}
\description{
gu function
}
