% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{hvals.lgcpPredict}
\alias{hvals.lgcpPredict}
\title{hvals.lgcpPredict function}
\usage{
\method{hvals}{lgcpPredict}(obj, ...)
}
\arguments{
\item{obj}{an object of class lgcpPredict}

\item{...}{additional arguments}
}
\value{
returns the values of h taken during the progress of the algorithm
}
\description{
Accessor function returning the value of \eqn{h}{h}, the MALA proposal scaling constant over the iterations of the algorithm for 
objects of class \code{lgcpPredict}
}
\seealso{
\link{lgcpPredict}
}
