% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{hvals}
\alias{hvals}
\title{hvals function}
\usage{
hvals(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{additional arguments}
}
\value{
method hvals
}
\description{
Generic function to return the values of the proposal scaling \eqn{h}{h} in the MCMC algorithm.
}
