% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{identifygrid}
\alias{identifygrid}
\title{identifygrid function}
\usage{
identifygrid(x, y)
}
\arguments{
\item{x}{the x grid centroids}

\item{y}{the y grid centroids}
}
\value{
a 2 x n matrix containing the grid indices of the points of interest, where n is the number of points selected via the mouse.
}
\description{
Identifies the indices of grid cells on plots of objects.
}
\seealso{
\link{lgcpPredict}, \link{loc2poly}, \link{identify.lgcpPredict}
}
