% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptiveSchemes.R
\name{initialiseAMCMC}
\alias{initialiseAMCMC}
\title{initialiseAMCMC function}
\usage{
initialiseAMCMC(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{additional arguments}
}
\value{
method intialiseAMCMC
}
\description{
A generic to be used for the purpose of user-defined adaptive MCMC schemes,
initialiseAMCMC tells the MALA algorithm which value of h to use first. See
lgcp vignette, code{vignette("lgcp")}, for further details on writing adaptive MCMC schemes.
}
\seealso{
\link{initialiseAMCMC.constanth}, \link{initialiseAMCMC.andrieuthomsh}
}
