% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stpppClassDef.R
\name{integerise}
\alias{integerise}
\title{integerise function}
\usage{
integerise(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{additional arguments}
}
\value{
method integerise
}
\description{
Generic function for converting the time variable of an stppp object.
}
\seealso{
\link{integerise.stppp}
}
