% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{intens}
\alias{intens}
\title{intens function}
\usage{
intens(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{additional arguments}
}
\value{
method intens
}
\description{
Generic function to return the Poisson Intensity.
}
\seealso{
\link{lgcpPredict}, \link{intens.lgcpPredict}
}
