% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterEstimation.R
\name{lambdaEst}
\alias{lambdaEst}
\title{lambdaEst function}
\usage{
lambdaEst(xyt, ...)
}
\arguments{
\item{xyt}{an object}

\item{...}{additional arguments}
}
\value{
method lambdaEst
}
\description{
Generic function for estimating bivariate densities by eye. Specific methods exist for stppp objects and ppp objects.
}
\seealso{
\link{lambdaEst.stppp}, \link{lambdaEst.ppp}
}
