% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{lgcpbayes}
\alias{lgcpbayes}
\title{lgcpbayes function}
\usage{
lgcpbayes()
}
\value{
displays the vignette by calling browseURL
}
\description{
Display the introductory vignette for the lgcp package.
}
