% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{lgcpgrid}
\alias{lgcpgrid}
\title{lgcpgrid function}
\usage{
lgcpgrid(grid, ...)
}
\arguments{
\item{grid}{a list object with each member of the list being a numeric matrix, each matrix having the same dimension}

\item{...}{other arguments}
}
\value{
method lgcpgrid
}
\description{
Generic function for the hadling of list objects where each element of the list is
a matrix. Each matrix is assumed to have the same dimension. Such objects arise from the 
various routines in the package lgcp.
}
\details{
lgcpgrid objects are list objects with names len, nrow, ncol, grid, xvals, yvals, zvals. The first three elements of the list
store the dimension of the object, the fourth element, grid, is itself a list object consisting of matrices in which the 
data is stored. The last three arguments can be used to give what is effectively a 3 dimensional array a physical reference.

For example, the mean of Y from a call to lgcpPredict, obj$y.mean for example, is stored in an lgcpgrid object. If several time points have been
stored in the call to lgcpPredict, then the grid element of the lgcpgrid object contains the output for each of the time points in succession. So 
the first element, obj$y.mean$grid[[1]],contains the output from the first time point and so on.
}
\seealso{
\link{lgcpgrid.list}, \link{lgcpgrid.array}, \link{lgcpgrid.matrix}
}
