% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{mcmctrace}
\alias{mcmctrace}
\title{mcmctrace function}
\usage{
mcmctrace(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{additional arguments}
}
\value{
method mcmctrace
}
\description{
Generic function to extract the information required to produce MCMC trace plots.
}
