% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{meanfield.lgcpPredict}
\alias{meanfield.lgcpPredict}
\title{meanfield.lgcpPredict function}
\usage{
\method{meanfield}{lgcpPredict}(obj, ...)
}
\arguments{
\item{obj}{an object of class lgcpPredict}

\item{...}{additional arguments}
}
\value{
returns the cell-wise mean of Y computed via Monte Carlo.
}
\description{
This is an accessor function for objects of class \code{lgcpPredict} and returns the mean of the 
field Y as an lgcpgrid object.
}
\seealso{
\link{lgcpPredict}, \link{lgcpgrid}
}
