% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcp-INLA.R
\name{meanfield.lgcpPredictINLA}
\alias{meanfield.lgcpPredictINLA}
\title{meanfield.lgcpPredictINLA function}
\usage{
\method{meanfield}{lgcpPredictINLA}(obj, ...)
}
\arguments{
\item{obj}{an object of class lgcpPredictINLA}

\item{...}{other arguments}
}
\value{
the mean of the latent field
}
\description{
A function to return the mean of the latent field from a call to lgcpPredictINLA output.
}
