% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstpppClassDef.R
\name{mstppp}
\alias{mstppp}
\title{mstppp function}
\usage{
mstppp(P, ...)
}
\arguments{
\item{P}{an object}

\item{...}{additional arguments}
}
\value{
method mstppp
}
\description{
Generic function used in the construction of marked space-time planar point patterns. An mstppp object is like an stppp object,
but with an extra component containing a data frame (the mark information).
}
\details{
Observations are assumed to occur in the plane and the observation window is assumed not to change over time.
}
\seealso{
\link{mstppp}, \link{mstppp.ppp}, \link{mstppp.list}
}
