% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{multiply.list}
\alias{multiply.list}
\title{multiply.list function}
\usage{
multiply.list(list1, list2)
}
\arguments{
\item{list1}{a list of objects that could be summed using "+"}

\item{list2}{a list of objects that could be summed using "+"}
}
\value{
a list with ith entry the sum of list1[[i]] and list2[[i]]
}
\description{
This function multiplies the elements of two list objects together and returns the result in another list object.
}
