% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcp-INLA.R
\name{neigh2D}
\alias{neigh2D}
\title{neigh2D function}
\usage{
neigh2D(i, j, ns, M, N)
}
\arguments{
\item{i}{cell index i}

\item{j}{cell index j}

\item{ns}{number of neighbours either side}

\item{M}{size of grid in x direction}

\item{N}{size of grid in y direction}
}
\value{
the cell indices of the neighbours
}
\description{
A function to compute the neighbours of a cell on a toral grid
}
