\name{nullAverage}
\alias{nullAverage}
\title{nullAverage function...}
\description{nullAverage function}
\details{A null scheme, that does not perform any computation in the running of \code{lgcpPredict}, it is the default
value of \code{gridmeans} in the argument \code{output.control}.}
\value{object of class nullAverage}
\seealso{\link{setoutput}, \link{lgcpPredict}, \link{GAinitialise}, \link{GAupdate}, \link{GAfinalise}, \link{GAreturnvalue}}
