% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridAverageMethods.R
\name{nullAverage}
\alias{nullAverage}
\title{nullAverage function}
\usage{
nullAverage()
}
\value{
object of class nullAverage
}
\description{
A null scheme, that does not perform any computation in the running of \code{lgcpPredict}, it is the default
value of \code{gridmeans} in the argument \code{output.control}.
}
\seealso{
\link{setoutput}, \link{lgcpPredict}, \link{GAinitialise}, \link{GAupdate}, \link{GAfinalise}, \link{GAreturnvalue}
}
