% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridFunctionMethods.R
\name{nullFunction}
\alias{nullFunction}
\title{nullFunction function}
\usage{
nullFunction()
}
\value{
object of class nullFunction
}
\description{
This is a null function and performs no action.
}
\seealso{
\link{setoutput}, \link{GFinitialise}, \link{GFupdate}, \link{GFfinalise}, \link{GFreturnvalue}
}
