% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{plotExceed}
\alias{plotExceed}
\title{plotExceed function}
\usage{
plotExceed(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{additional arguments}
}
\value{
generic function returning method plotExceed
}
\description{
A generic function for plotting exceedance probabilities.
}
\seealso{
\link{plotExceed.lgcpPredict}, \link{plotExceed.array}
}
