% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridFunctionMethods.R
\name{print.dump2dir}
\alias{print.dump2dir}
\title{print.dump2dir function}
\usage{
\method{print}{dump2dir}(x, ...)
}
\arguments{
\item{x}{an object of class dump2dir}

\item{...}{additional arguments}
}
\value{
nothing
}
\description{
Display function for dump2dir objects.
}
\seealso{
\link{dump2dir},
}
