% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAtRiskClassDef.R
\name{print.fromFunction}
\alias{print.fromFunction}
\title{print.fromFunction function}
\usage{
\method{print}{fromFunction}(x, ...)
}
\arguments{
\item{x}{an object of class spatialAtRisk}

\item{...}{additional arguments}
}
\value{
prints the object
}
\description{
Print method for objects of class fromFunction.
}
