% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAtRiskClassDef.R
\name{print.fromSPDF}
\alias{print.fromSPDF}
\title{print.fromSPDF function}
\usage{
\method{print}{fromSPDF}(x, ...)
}
\arguments{
\item{x}{an object of class spatialAtRisk}

\item{...}{additional arguments}
}
\value{
prints the object
}
\description{
Print method for objects of class fromSPDF.
}
