% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{print.gridaverage}
\alias{print.gridaverage}
\title{print.gridaverage function}
\usage{
\method{print}{gridaverage}(x, ...)
}
\arguments{
\item{x}{an object of class gridaverage}

\item{...}{other arguments}
}
\value{
just prints out details
}
\description{
Print method for \code{gridaverage} objects
}
