% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{print.lgcpgrid}
\alias{print.lgcpgrid}
\title{print.lgcpgrid function}
\usage{
\method{print}{lgcpgrid}(x, ...)
}
\arguments{
\item{x}{an object of class lgcpgrid}

\item{...}{other arguments}
}
\value{
just prints out details to the console
}
\description{
Print method for lgcp grid objects.
}
\seealso{
\link{lgcpgrid.list}, \link{lgcpgrid.array}, \link{as.list.lgcpgrid},
\link{summary.lgcpgrid} \link{quantile.lgcpgrid} \link{image.lgcpgrid} \link{plot.lgcpgrid}
}
