\name{print.lgcpPredict}
\alias{print.lgcpPredict}
\title{print.lgcpPredict function}
\usage{
  \method{print}{lgcpPredict} (x, ...)
}
\arguments{
  \item{x}{an object of class lgcpPredict}

  \item{...}{additional arguments}
}
\value{
  just prints information to the screen
}
\description{
  Print method for lgcpPredict objects.
}
\seealso{
  \link{lgcpPredict}
}

