% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arealMethods.R
\name{print.stapp}
\alias{print.stapp}
\title{print.stapp function}
\usage{
\method{print}{stapp}(x, printhead = TRUE, ...)
}
\arguments{
\item{x}{an object of class stapp}

\item{printhead}{whether or not to print the head of the counts matrix}

\item{...}{additional arguments}
}
\value{
prints the stapp object x
}
\description{
Print method for stapp objects
}
