\name{print.temporalAtRisk}
\alias{print.temporalAtRisk}
\title{print...}
\usage{\method{print}{temporalAtRisk} (x, ...)}
\description{print.temporalAtRisk function}
\details{Printing method for temporalAtRisk objects.}
\value{print the object}
\seealso{\link{temporalAtRisk}, \link{spatialAtRisk}, \link{temporalAtRisk.numeric}, \link{temporalAtRisk.function}, \link{constantInTime}, \link{constantInTime.numeric}, \link{constantInTime.stppp}, \link{plot.temporalAtRisk}}
\arguments{\item{x}{an object}
\item{...}{additional arguments}}
