% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpStructures.R
\name{PriorSpec}
\alias{PriorSpec}
\title{PriorSpec function}
\usage{
PriorSpec(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{additional arguments}
}
\value{
method PriorSpec
}
\description{
Generic for declaring that an object is of valid type for use as as prior in lgcp. For further details and examples, see the vignette "Bayesian_lgcp".
}
\seealso{
\link{PriorSpec.list}
}
