% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstpppClassDef.R
\name{rescale.mstppp}
\alias{rescale.mstppp}
\title{rescale.mstppp function}
\usage{
\method{rescale}{mstppp}(X, s, unitname)
}
\arguments{
\item{X}{an object of class mstppp}

\item{s}{scale as in rescale.ppp: x and y coordinaes are scaled by 1/s}

\item{unitname}{parameter as defined in ?rescale}
}
\value{
a ppp object without observation times
}
\description{
Rescale an mstppp object. Similar to rescale.ppp
}
