\name{rescale.stppp}
\alias{rescale.stppp}
\title{rescale.stppp function}
\usage{
\method{rescale}{stppp}(X, s)
}
\arguments{
  \item{X}{an object of class stppp}

  \item{s}{scale as in rescale.ppp: x and y coordinaes are
  scaled by 1/s}
}
\value{
a ppp object without observation times
}
\description{
Rescale an stppp object. Similar to rescale.ppp
}

