% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectObsWindow.R
\name{roteffgain}
\alias{roteffgain}
\title{roteffgain function}
\usage{
roteffgain(xyt, cellwidth)
}
\arguments{
\item{xyt}{an object of class stppp}

\item{cellwidth}{size of grid on which to do MALA}
}
\value{
whether or not there woud be any efficiency gain in the MALA by rotating window
}
\description{
Compute whether there might be any advantage in rotating the observation window in the object xyt for a proposed cell width.
}
\seealso{
\link{getRotation.stppp}
}
