\name{rotmat}
\alias{rotmat}
\title{rotmat function}
\usage{
rotmat(theta)
}
\arguments{
  \item{theta}{an angle in radians}
}
\value{
the transformation matrix corresponding to an anticlockwise
rotation of theta radians about the origin
}
\description{
This function returns a rotation matrix corresponding to an
anticlockwise rotation of theta radians about the origin
}

