% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{rr}
\alias{rr}
\title{rr function}
\usage{
rr(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{additional arguments}
}
\value{
method rr
}
\description{
Generic function to return relative risk.
}
\seealso{
\link{lgcpPredict}, \link{rr.lgcpPredict}
}
