\name{seintens}
\alias{seintens}
\title{seintens function}
\usage{
  seintens(obj, ...)
}
\arguments{
  \item{obj}{an object}

  \item{...}{additional arguments}
}
\value{
  method seintens
}
\description{
  Generic function to return the standard error of the
  Poisson Intensity.
}
\seealso{
  \link{lgcpPredict}, \link{seintens.lgcpPredict}
}

