\name{selectObsWindow.default}
\alias{selectObsWindow.default}
\title{selectObsWindow...}
\usage{\method{selectObsWindow}{default} (xyt, ...)}
\description{selectObsWindow.default function}
\details{Default method, note at present, there is only an implementation for stppp objects.}
\value{currently no default implementation}
\arguments{\item{xyt}{an object}
\item{...}{additional arguments}}
