\name{selectObsWindow}
\alias{selectObsWindow}
\title{selectObsWindow function...}
\usage{selectObsWindow(xyt, ...)}
\description{selectObsWindow function}
\details{See ?selectObsWindow.stppp for further details on usage. This is a generic function for the purpose of selecting an observation window 
(or more precisely a bounding box) to contain the extended FFT grid.}
\value{method selectObsWindow}
\seealso{\link{selectObsWindow.default}, \link{selectObsWindow.stppp}}
\arguments{\item{xyt}{an object}
\item{...}{additional arguments}}
