\name{selectObsWindow}
\alias{selectObsWindow}
\title{selectObsWindow function}
\usage{
  selectObsWindow(xyt, ...)
}
\arguments{
  \item{xyt}{an object}

  \item{...}{additional arguments}
}
\value{
  method selectObsWindow
}
\description{
  See ?selectObsWindow.stppp for further details on usage.
  This is a generic function for the purpose of selecting
  an observation window (or more precisely a bounding box)
  to contain the extended FFT grid.
}
\seealso{
  \link{selectObsWindow.default},
  \link{selectObsWindow.stppp}
}

