% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectObsWindow.R
\name{selectObsWindow.stppp}
\alias{selectObsWindow.stppp}
\title{selectObsWindow.stppp function}
\usage{
\method{selectObsWindow}{stppp}(xyt, cellwidth, ...)
}
\arguments{
\item{xyt}{an object of class stppp}

\item{cellwidth}{size of the grid spacing in chosen units (equivalent to the cell width argument in \link{lgcpPredict})}

\item{...}{additional arguments}
}
\value{
a resized stppp object together with grid sizes M and N ready for FFT, together with the FFT grid locations, can be useful for estimating lambda(s)
}
\description{
This function computes an appropriate observation window on which to perform prediction. Since the FFT grid
must have dimension 2^M by 2^N for some M and N, the window \code{xyt$window}, is extended to allow this to be fit in for a given cell width.
}
\details{
!!NOTE!! that this function also returns the grid ($xvals and $yvals) on which the FFT (and hence MALA) will be performed. It is useful to
define spatialAtRiskobjects on this grid to prevent loss of information from the bilinear interpolation that takes place as part of the fitting 
algorithm.
}
\seealso{
\link{spatialAtRisk}
}
