% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{serr.lgcpPredict}
\alias{serr.lgcpPredict}
\title{serr.lgcpPredict function}
\usage{
\method{serr}{lgcpPredict}(obj, ...)
}
\arguments{
\item{obj}{an lgcpPredict object}

\item{...}{additional arguments}
}
\value{
Standard error of the relative risk as computed by MCMC.
}
\description{
Accessor function returning the standard error of relative risk as an lgcpgrid object.
}
\seealso{
\link{lgcpPredict}
}
