% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{serr}
\alias{serr}
\title{serr function}
\usage{
serr(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{additional arguments}
}
\value{
method serr
}
\description{
Generic function to return standard error of relative risk.
}
\seealso{
\link{lgcpPredict}, \link{serr.lgcpPredict}
}
