% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setoutput.R
\name{setoutput}
\alias{setoutput}
\title{setoutput function}
\usage{
setoutput(gridfunction = NULL, gridmeans = NULL)
}
\arguments{
\item{gridfunction}{what to do with the latent field, but default this set to nothing, but could save output to a directory, see ?dump2dir}

\item{gridmeans}{list of Monte Carlo averages to compute, see ?MonteCarloAverage}
}
\value{
output parameters
}
\description{
Sets output functionality for \link{lgcpPredict} via the main functions \link{dump2dir} and \link{MonteCarloAverage}. Note that it is possible for
the user to create their own \code{gridfunction} and \code{gridmeans} schemes.
}
\seealso{
\link{lgcpPredict}, \link{dump2dir}, \link{MonteCarloAverage}
}
