% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txtProgressBar2.R
\name{setTxtProgressBar2}
\alias{setTxtProgressBar2}
\title{set the progress bar}
\usage{
setTxtProgressBar2(pb, value, title = NULL, label = NULL)
}
\arguments{
\item{pb}{text progress bar object}

\item{value}{new value}

\item{title}{ignored}

\item{label}{text for end of progress bar}
}
\description{
update a text progress bar. See help(txtProgressBar) for more info.
}
