% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{showGrid.default}
\alias{showGrid.default}
\title{showGrid.default function}
\usage{
\method{showGrid}{default}(x, y, ...)
}
\arguments{
\item{x}{an vector of grid values for the x coordinates}

\item{y}{an vector of grid values for the y coordinates}

\item{...}{additional arguments passed to points}
}
\value{
plots grid centroids on the current graphics device
}
\description{
Default method for printing a grid to a screen. Arguments are vectors giving the x any y coordinates of the
centroids.
}
\seealso{
\link{showGrid.lgcpPredict}, \link{showGrid.stppp}
}
