\name{showGrid}
\alias{showGrid}
\title{showGrid function...}
\usage{showGrid(x, ...)}
\description{showGrid function}
\details{Generic method for displaying the FFT grid used in computation.}
\value{generic function returning method showGrid}
\seealso{\link{showGrid.default}, \link{showGrid.lgcpPredict}, \link{showGrid.stppp}}
\arguments{\item{x}{an object}
\item{...}{additional arguments}}
