% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{showGrid.stppp}
\alias{showGrid.stppp}
\title{showGrid.stppp function}
\usage{
\method{showGrid}{stppp}(x, ...)
}
\arguments{
\item{x}{an object of class stppp. Note this function oly applies to SIMULATED data.}

\item{...}{additional arguments  passed to points}
}
\value{
plots grid centroids on the current graphics device. FOR SIMULATED DATA ONLY.
}
\description{
If an stppp object has been created via simulation, ie using the function \code{lgcpSim}, then
this function will display the grid centroids that were used in the simulation
}
\examples{
\dontrun{xyt <- lgcpSim()}
\dontrun{plot(xyt)}
\dontrun{showGrid(xyt,pch="+",cex=0.5)}
}
\seealso{
\link{lgcpSim}, \link{showGrid.default}, \link{showGrid.lgcpPredict}
}
