% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcp-INLA.R
\name{sparsebase}
\alias{sparsebase}
\title{sparsebase function}
\usage{
sparsebase(base)
}
\arguments{
\item{base}{base matrix of a block circulant matrix}
}
\value{
...
}
\description{
A function that returns the full precision matrix in sparse format from the base of a block circulant matrix, see ?Matrix::sparseMatrix
}
