% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAtRiskClassDef.R
\name{spatialIntensities}
\alias{spatialIntensities}
\title{spatialIntensities function}
\usage{
spatialIntensities(X, ...)
}
\arguments{
\item{X}{an object}

\item{...}{additional arguments}
}
\value{
method spatialintensities
}
\description{
Generic method for extracting spatial intensities.
}
\seealso{
\link{spatialIntensities.fromXYZ}, \link{spatialIntensities.fromSPDF}
}
