% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptiveSchemes.R
\name{updateAMCMC}
\alias{updateAMCMC}
\title{updateAMCMC function}
\usage{
updateAMCMC(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{additional arguments}
}
\value{
method updateAMCMC
}
\description{
A generic to be used for the purpose of user-defined adaptive MCMC schemes,
updateAMCMC tells the MALA algorithm how to update the value of h. See
lgcp vignette, code{vignette("lgcp")}, for further details on writing adaptive MCMC schemes.
}
\seealso{
\link{updateAMCMC.constanth}, \link{updateAMCMC.andrieuthomsh}
}
