% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{varfield.lgcpPredict}
\alias{varfield.lgcpPredict}
\title{varfield.lgcpPredict function}
\usage{
\method{varfield}{lgcpPredict}(obj, ...)
}
\arguments{
\item{obj}{an object of class lgcpPredict}

\item{...}{additional arguments}
}
\value{
returns the cell-wise variance of Y computed via Monte Carlo.
}
\description{
This is an accessor function for objects of class \code{lgcpPredict} and returns the variance of the 
field Y as an lgcpgrid object.
}
\seealso{
\link{lgcpPredict}
}
