% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcp-INLA.R
\name{varfield.lgcpPredictINLA}
\alias{varfield.lgcpPredictINLA}
\title{varfield.lgcpPredictINLA function}
\usage{
\method{varfield}{lgcpPredictINLA}(obj, ...)
}
\arguments{
\item{obj}{an object of class lgcpPredictINLA}

\item{...}{other arguments}
}
\value{
the variance of the latent field
}
\description{
A function to return the variance of the latent field from a call to lgcpPredictINLA output.
}
