\name{varfield}
\alias{varfield}
\title{varfield function...}
\usage{varfield(obj, ...)}
\description{varfield function}
\details{Generic function to extract the variance of the latent field Y.}
\value{method meanfield}
\seealso{\link{lgcpPredict}}
\arguments{\item{obj}{an object}
\item{...}{additional arguments}}
