% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{varfield}
\alias{varfield}
\title{varfield function}
\usage{
varfield(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{additional arguments}
}
\value{
method meanfield
}
\description{
Generic function to extract the variance of the latent field Y.
}
\seealso{
\link{lgcpPredict}
}
