% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpMethods.R
\name{window.lgcpPredict}
\alias{window.lgcpPredict}
\title{window.lgcpPredict function}
\usage{
\method{window}{lgcpPredict}(x, ...)
}
\arguments{
\item{x}{an object of class lgcpPredict}

\item{...}{additional arguments}
}
\value{
returns the observation window used durign computation
}
\description{
Accessor function returning the observation window from objects of class \code{lgcpPredict}. Note that for
computational purposes, the window of an \code{stppp} object will be extended to accommodate the requirement that 
the dimensions must be powers of 2. The function \code{window.lgcpPredict} returns the extended window.
}
\seealso{
\link{lgcpPredict}
}
