\name{xvals.fromXYZ}
\alias{xvals.fromXYZ}
\title{xvals.fromXYZ function}
\usage{
\method{xvals}{fromXYZ}(obj, ...)
}
\arguments{
  \item{obj}{a spatialAtRisk object}

  \item{...}{additional arguments}
}
\value{
the x values
}
\description{
Method for extracting 'x values' from an object of class
fromXYZ
}
\seealso{
\link{xvals}, \link{yvals}, \link{zvals},
\link{xvals.default}, \link{yvals.default},
\link{zvals.default}, \link{yvals.fromXYZ},
\link{zvals.fromXYZ}, \link{xvals.SpatialGridDataFrame},
\link{yvals.SpatialGridDataFrame},
\link{zvals.SpatialGridDataFrame}
}

