% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAtRiskClassDef.R
\name{xvals}
\alias{xvals}
\title{xvals function}
\usage{
xvals(obj, ...)
}
\arguments{
\item{obj}{an object of class spatialAtRisk}

\item{...}{additional arguments}
}
\value{
the xvals method
}
\description{
Generic for extractign the 'x values' from an object.
}
\seealso{
\link{yvals}, \link{zvals}, \link{xvals.default}, \link{yvals.default}, \link{zvals.default}, \link{xvals.fromXYZ}, \link{yvals.fromXYZ}, \link{zvals.fromXYZ}, \link{xvals.SpatialGridDataFrame}, \link{yvals.SpatialGridDataFrame}, \link{zvals.SpatialGridDataFrame}
}
