% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAtRiskClassDef.R
\name{xvals.SpatialGridDataFrame}
\alias{xvals.SpatialGridDataFrame}
\title{xvals.SpatialGridDataFrame function}
\usage{
\method{xvals}{SpatialGridDataFrame}(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{additional arguments}
}
\value{
the x values
}
\description{
Method for extracting 'x values' from an object of class spatialGridDataFrame
}
\seealso{
\link{xvals},  \link{yvals}, \link{zvals}, \link{xvals.default}, \link{yvals.default}, \link{zvals.default}, \link{xvals.fromXYZ}, \link{yvals.fromXYZ}, \link{zvals.fromXYZ}, \link{yvals.SpatialGridDataFrame}, \link{zvals.SpatialGridDataFrame}
}
