% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAtRiskClassDef.R
\name{yvals}
\alias{yvals}
\title{yvals function}
\usage{
yvals(obj, ...)
}
\arguments{
\item{obj}{an object of class spatialAtRisk}

\item{...}{additional arguments}
}
\value{
the yvals method
}
\description{
Generic for extractign the 'y values' from an object.
}
\seealso{
\link{xvals}, \link{zvals}, \link{xvals.default}, \link{yvals.default}, \link{zvals.default}, \link{xvals.fromXYZ}, \link{yvals.fromXYZ}, \link{zvals.fromXYZ}, \link{xvals.SpatialGridDataFrame}, \link{yvals.SpatialGridDataFrame}, \link{zvals.SpatialGridDataFrame}
}
