% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAtRiskClassDef.R
\name{zvals.fromXYZ}
\alias{zvals.fromXYZ}
\title{zvals.fromXYZ function}
\usage{
\method{zvals}{fromXYZ}(obj, ...)
}
\arguments{
\item{obj}{a spatialAtRisk object}

\item{...}{additional arguments}
}
\value{
the z values
}
\description{
Method for extracting 'z values' from an object of class fromXYZ
}
\seealso{
\link{xvals},  \link{yvals}, \link{zvals}, \link{xvals.default}, \link{yvals.default}, \link{zvals.default}, \link{xvals.fromXYZ}, \link{yvals.fromXYZ}, \link{xvals.SpatialGridDataFrame}, \link{yvals.SpatialGridDataFrame}, \link{zvals.SpatialGridDataFrame}
}
