% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAtRiskClassDef.R
\name{zvals}
\alias{zvals}
\title{zvals function}
\usage{
zvals(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{additional arguments}
}
\value{
the zvals method
}
\description{
Generic for extractign the 'z values' from an object.
}
\seealso{
\link{xvals},  \link{yvals}, \link{xvals.default}, \link{yvals.default}, \link{zvals.default}, \link{xvals.fromXYZ}, \link{yvals.fromXYZ}, \link{zvals.fromXYZ}, \link{xvals.SpatialGridDataFrame}, \link{yvals.SpatialGridDataFrame}, \link{zvals.SpatialGridDataFrame}
}
