\name{zvals.SpatialGridDataFrame}
\alias{zvals.SpatialGridDataFrame}
\title{zvals.SpatialGridDataFrame function}
\usage{
  \method{zvals}{SpatialGridDataFrame} (obj, ...)
}
\arguments{
  \item{obj}{an object}

  \item{...}{additional arguments}
}
\value{
  the z values
}
\description{
  Method for extracting 'z values' from an object of class
  SpatialGridDataFrame
}
\seealso{
  \link{xvals}, \link{yvals}, \link{zvals},
  \link{xvals.default}, \link{yvals.default},
  \link{zvals.default}, \link{xvals.fromXYZ},
  \link{yvals.fromXYZ}, \link{zvals.fromXYZ},
  \link{xvals.SpatialGridDataFrame},
  \link{yvals.SpatialGridDataFrame}
}

